% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lat_bands.R
\name{add_lat_bands}
\alias{add_lat_bands}
\title{A function to categorize latitudes}
\usage{
add_lat_bands(data, tropic_cut = 30, mid_cut = 70)
}
\arguments{
\item{data}{An input data frame with a column, latcol, of latitude data to be categorized into tropic, mid, and high}

\item{tropic_cut}{The boundary between tropic and mid latitudes, defaults to 30}

\item{mid_cut}{The boundary between mid and high latitudes, defaults to 70}
}
\description{
A function to categorize latitudes
}
