% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{age_cohort}
\alias{age_cohort}
\title{Age up all living agents to the current timestep}
\usage{
age_cohort(cohort, k)
}
\arguments{
\item{cohort}{Population data frame}

\item{k}{Current timestep (age to assign)}
}
\value{
Updated cohort data frame
}
\description{
Age up all living agents to the current timestep
}
\keyword{internal}
