% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aicc.R
\name{aicc}
\alias{aicc}
\alias{aicc,ProgressBM-method}
\alias{aicc,ProgressCLM-method}
\alias{aicc,ProgressBMM-method}
\alias{aicc,ProgressCLMM-method}
\alias{aicc,ProgressModelList-method}
\title{Akaike Information Criterion (corrected)}
\usage{
aicc(object, ...)

\S4method{aicc}{ProgressBM}(object, ...)

\S4method{aicc}{ProgressCLM}(object, ...)

\S4method{aicc}{ProgressBMM}(object, ...)

\S4method{aicc}{ProgressCLMM}(object, ...)

\S4method{aicc}{ProgressModelList}(object, ...)
}
\arguments{
\item{object}{an object of class \code{ProgressModel} or \code{ProgressModelList}.}

\item{...}{extra arguments.}
}
\value{
numeric. The model AICc.
}
\description{
Akaike Information Criterion (corrected)
}
\examples{
\dontrun{
# Create a Region object
library(cronus)
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))

# Create a model
object1 <- new("ProgressBM",
               region = region,
               crop = "Corn",
               data = data_progress$Corn,
               formula = "CumPercentage ~ Time + agdd") # ProgressModel

# Create another model
object2 <- new("ProgressCLM",
               region = region,
               crop = "Soybeans",
               data = data_progress$Soybeans,
               formula = "Stage ~ Time + agdd + adayl") # ProgressModel

# Concatenate the models
object <- c(object1, object2) # ProgressModelList

# Fit
object <- fit(object)

# Plot
plot(object, cumulative = TRUE, seasons = 2002)

# Predict
predict(object, data_progress)

# Evaluate
object <- evaluate(object, maxsam = 100, seed = 1)
plot_metrics(object)

# Summarize
summary(object)

# Report
report(object, name = "example_report", path = getwd())
}
}
