% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{armspe}
\alias{armspe}
\title{Average Root Mean Square Prediction Error}
\usage{
armspe(data)
}
\arguments{
\item{data}{data.frame. A data.frame with columns Stage, Time and spe.}
}
\value{
A data.frame with columns Stage, Time and rmspe.
}
\description{
Average Root Mean Square Prediction Error
}
