% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/siler_helper_functions.R
\name{baseline_death_prob}
\alias{baseline_death_prob}
\title{Compute baseline Siler death probability}
\usage{
baseline_death_prob(age, regime)
}
\arguments{
\item{age}{Numeric. Age in years.}

\item{regime}{Data frame with Siler parameters (a1, b1, a2, a3, b3).}
}
\value{
Numeric death probability.
}
\description{
Compute baseline Siler death probability
}
