% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{c,PersephoneModel-method}
\alias{c,PersephoneModel-method}
\title{Concatenate}
\usage{
\S4method{c}{PersephoneModel}(x, ...)
}
\arguments{
\item{x}{PersephoneModel.}

\item{...}{extra models to concatenate.}
}
\value{
PersephoneModelList
}
\description{
Concatenate PersephoneModel objects into a PersephoneModelList.
}
\examples{
\dontrun{
object1 <- new("PersephoneCumLink")
object2 <- new("PersephoneBinomial")
object <- c(object1, object2)
}
}
