% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CoaleDemenyWest}
\alias{CoaleDemenyWest}
\alias{CoaleDemenyWestF3}
\alias{CoaleDemenyWestF5}
\alias{CoaleDemenyWestF11}
\alias{CoaleDemenyWestF15}
\alias{CoaleDemenyWestF17}
\alias{CoaleDemenyWestF21}
\title{Coale-Demeny West Female Model Life Table Siler Parameters}
\format{
A data frame with 1 row and 6 columns.

An object of class \code{data.frame} with 1 rows and 6 columns.

An object of class \code{data.frame} with 1 rows and 6 columns.

An object of class \code{data.frame} with 1 rows and 6 columns.

An object of class \code{data.frame} with 1 rows and 6 columns.

An object of class \code{data.frame} with 1 rows and 6 columns.

An object of class \code{data.frame} with 1 rows and 6 columns.
}
\usage{
CoaleDemenyWestF3

CoaleDemenyWestF5

CoaleDemenyWestF11

CoaleDemenyWestF15

CoaleDemenyWestF17

CoaleDemenyWestF21
}
\description{
Siler function parameters fit to Coale & Demeny West model life tables
for females. Each object is a one-row data frame with columns:
a1, b1, a2, a3, b3, name.
}
\keyword{datasets}
