% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ctw.R
\name{convert_ctw}
\alias{convert_ctw}
\title{A function to convert input changes in T, changes in W, and values of C to be standardized with C3MP standaradization}
\usage{
convert_ctw(ctwdataframe)
}
\arguments{
\item{ctwdataframe}{A dataframe of C, T, W values to be converted}
}
\description{
A function to convert input changes in T, changes in W, and values of C to be standardized with C3MP standaradization
}
