% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{create}
\alias{create}
\title{Create new objects}
\usage{
create(prm)
}
\arguments{
\item{prm}{a list with all the necessary elements to build a model.}
}
\value{
An S4 object of class \code{ProgressModelList}
}
\description{
The function creates a \code{ProgressModelList}.
}
\details{
The function uses the first entry of each list element to create
the first \code{ProgressModel}, then the second entry and so on.
}
\examples{
\dontrun{
# Create a Region object
library(cronus)
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))

# Class
Class <- "ProgressBM"

# Formula
formula <- list(c(global = "CumPercentage ~ Time + agdd * adayl"),
                c(base = "CumPercentage ~ Time + agdd"))

# Link
link <- list("logit", "probit")

# Set the parameters
prm <- combine(Class = Class,
               formula = formula,
               link = link,
               region = list(region),
               data = data_progress)

prm$crop <- names(prm$data)

# Create the object
object <- create(prm)
}
}
