% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{create_cohort}
\alias{create_cohort}
\title{Create the initial cohort data frame}
\usage{
create_cohort(cohort_size)
}
\arguments{
\item{cohort_size}{Number of agents in the starting cohort}
}
\value{
A data frame with columns: agent_id, age, lesion, dead, in_sample
}
\description{
Create the initial cohort data frame
}
\keyword{internal}
