% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{create_title}
\alias{create_title}
\title{Create plot title}
\usage{
create_title(crop, seasons)
}
\arguments{
\item{crop}{character. A crop of interest.}

\item{seasons}{numeric. A vector of seasons.}
}
\value{
character. A title for the plot
}
\description{
Create plot title
}
\examples{
\dontrun{
create_title("Corn", 2002)
create_title("Corn", 2002:2009)
}
}
