% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_progress}
\alias{data_progress}
\title{Corn and Soybeans progress data for Nebraska}
\format{
\subsection{\code{data_progress}}{

An object of class \code{ProgressList}, a list with two named elements, \code{Corn} and
\code{Soybeans}, of class \code{Progress}. Each one is a data.frame with columns:
\describe{
\item{Crop}{Crop name}
\item{Stage}{Phenological stage}
\item{Season}{Cultivation season / year}
\item{Time}{Integer time-step within season}
\item{Date}{Full date}
\item{CumPercentage}{Accumulated crop progress percentages}
\item{Percentage}{Crop progress percentages}
\item{agdd}{Accumulated growing degree days}
\item{adayl}{Accumulated daylight}
\item{aprcp}{Accumulated precipitation}
\item{aswe}{Accumulated snow vater equivalent}
\item{avp}{Accumulated vapor pressure}
}
}
}
\source{
\url{https://quickstats.nass.usda.gov/}
}
\usage{
data_progress
}
\description{
A subset of data from the USDA NASS Quick Stats crop progress product. The
dataset concerns 2 crops, corn and soybeans, for 20 years, from 2002 to 2021.
}
\keyword{datasets}
