% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eaDist_exact.R
\docType{data}
\name{eaDist_exact}
\alias{eaDist_exact}
\title{Distribution of the dates for Easter (Exact)}
\format{
Theoretical Distribution of Easter Dates over the span of 5.7 mio years
}
\usage{
data(eaDist_exact)
}
\description{
The date of Easter (Sunday) varies between March 22nd and April 25th.
The sequence of Easter is periodic with a period of 5.7 mio years.
}
\references{
Extracted on	15.02.20 from \url{https://ec.europa.eu/eurostat/cros/system/files/jd__calendars_0.docx}
}
\keyword{datasets}
