% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_func_form_ensemble.R
\name{fit_func_form_ensemble}
\alias{fit_func_form_ensemble}
\title{A function to fit the Bayesian model and write the resulting rethinking model,
point estimates and the standard deviation of the coefficients for mu,
point estimates and the standard deviation of the coefficients for sigma,
and a plot of the above data.}
\usage{
fit_func_form_ensemble(ensSiteData, testnmes, ctwtests,
  individSiteDataPath = "data/c3mpdata/csvs/",
  coefpath = "data/emulatorcoefficients", functionalform, sdfunctionalform,
  figurepath = "figures/bayeseval", ensemblename)
}
\arguments{
\item{ensSiteData}{A list of sites for ensemble under consideration}

\item{testnmes}{The names of the CTW tests}

\item{ctwtests}{The actual ctw tests in full}

\item{coefpath}{The location to write the outputs, defaults to "data/emulatorcoefficients"}

\item{functionalform}{The name functional form for the mean of the likelihood}

\item{sdfunctionalform}{The name of the functional form for the standard deviation of the likelihood}

\item{figurepath}{The location to write the plots of the coefficients + uncertainty, defaults to "figures/bayeseval"}

\item{ensemblename}{The name of the ensemble under consideration}

\item{individuSiteDataPath}{The location of the raw C3MP yield data, defaults to "data/c3mpdata/csvs/"}
}
\description{
A function to fit the Bayesian model and write the resulting rethinking model,
point estimates and the standard deviation of the coefficients for mu,
point estimates and the standard deviation of the coefficients for sigma,
and a plot of the above data.
}
