% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{form_lesion}
\alias{form_lesion}
\title{Roll for lesion formation for a single agent}
\usage{
form_lesion(
  cohort,
  i,
  formation_window_opens,
  formation_window_closes,
  lesion_formation_rate
)
}
\arguments{
\item{cohort}{Population data frame}

\item{i}{Index of the agent}

\item{formation_window_opens}{Age at which lesions can start forming}

\item{formation_window_closes}{Age at which lesions stop forming}

\item{lesion_formation_rate}{Annual probability of lesion formation}
}
\value{
Updated cohort data frame
}
\description{
Roll for lesion formation for a single agent
}
\keyword{internal}
