% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func_form_evaluate_plusstddev.R
\name{func_form_evaluate}
\alias{func_form_evaluate}
\title{A function to evaluate yield changes in response to input CTW values,
given a specific functional form for a specific ensemble
and the corresponding coefficients for mu and sigma.}
\usage{
func_form_evaluate(coeff, SDcoeff, ctwvalues, baseC = 360, baseT = 0,
  baseP = 1, functionalform, sdfunctionalform, ensembleName, stddev = TRUE,
  formaps = FALSE)
}
\arguments{
\item{coeff}{The point values of coefficients for mu}

\item{SDcoeff}{The point values of coefficients for sigma}

\item{ctwvalues}{The CTW values at which to evaluate}

\item{baseC}{The baseline value of CO2 concentration, defaults to C3MP baseline of 360 ppm}

\item{baseT}{The baseline value of temperature changes, defaults to C3MP baseline of 0C change}

\item{functionalform}{The name functional form for the mean of the likelihood}

\item{sdfunctionalform}{The name of the functional form for the standard deviation of the likelihood}

\item{ensembleName}{The name of the ensemble under consideration}

\item{stddev}{Whether to return the standard deviation of the yield estimate (ie whether to also evaluate sigma), defaults to TRUE}

\item{formaps}{Whether the outputs will be used for plotting maps, defaults to TRUE. Just changes some formatting.}

\item{base}{P The baseline value of precipitation changes, defaults to C3MP baseline of 0% change}
}
\description{
A function to evaluate yield changes in response to input CTW values,
given a specific functional form for a specific ensemble
and the corresponding coefficients for mu and sigma.
}
