% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep.R
\name{generate_output_directory}
\alias{generate_output_directory}
\title{Generate output directory path from parameters}
\usage{
generate_output_directory(params)
}
\arguments{
\item{params}{Named list of simulation parameters}
}
\value{
Character string directory path
}
\description{
Generate output directory path from parameters
}
