% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_qr_table.R
\name{generateQrTable}
\alias{generateQrTable}
\title{Generate figures from the EUROSTAT quality report}
\usage{
generateQrTable(x)
}
\arguments{
\item{x}{a persephone object}
}
\description{
This function uses the method \code{iterate} to create figures for a quality
report for each component of a hierarchical \code{persephone} object
}
\section{Columns}{

\itemize{
\item \strong{component} is a column added by \code{iterate()} and informs where in the
hierarchy a certain component is.
\item \strong{Method} is \code{TS} for \code{\link[=tramoseats]{tramoseats()}} and \code{x13} for \code{\link[=x13]{x13()}}
\item \strong{Period} is \code{12} for yearly measurements and \code{4} for quarterly
measurements
\item \strong{Start} denotes the start of the time series
\item \strong{End} denotes the end of the time series
\item \strong{Log.Transformation}
\item \strong{ARIMA.Model}
\item \strong{LeapYear}
\item \strong{MovingHoliday}
\item \strong{NbTF}
\item \strong{Noutliers} Number of outliers
\item \strong{Outlier1}
\item \strong{Outlier2}
\item \strong{Outlier3}
\item \strong{CombinedTest_SI}
\item \strong{Residual.Seasonality}
\item \strong{Residual.TD.Effect}
\item \strong{Q.Stat}
\item \strong{Final.Henderson.Filter}
\item \strong{Stage.2.Henderson.Filter}
\item \strong{Seasonal.Filter}
\item \strong{Max.Adj}
}
}

\examples{
obj_x13 <- per_x13(AirPassengers, "RSA3")
ht <- per_hts(a = obj_x13, b = obj_x13, method = "x13")
ht$run()
generate_qr_table(ht)

}
