% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTd.R
\name{genTd}
\alias{genTd}
\title{Generation of centered trading day regressor}
\usage{
genTd(ff = 12, fYear = 1960, lYear = 2099, ft, ww, reg = NULL)
}
\arguments{
\item{ff}{frequency of series}

\item{fYear}{first year of calculation of trading day regressor}

\item{lYear}{last year of calculation of trading day regressor}

\item{ft}{holidays, list of 1. exact date ("mm-dd") and/or 2. name of day
(listHolidays() from timeDate function) and/or 3. easter relation
(e.g. "easter+39")}

\item{ww}{vector of individual weights for each holiday with length of ft}

\item{reg}{seven day of a week regressor (reg = "td") or weekday/weekend
regressor (reg = "wd"). Not yet implemented.}
}
\value{
matrix of centered trading day variables
}
\description{
Generation of centered trading day regressor
}
