% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_class}
\alias{get_class}
\alias{get_class,PersephoneModel-method}
\alias{get_class,PersephoneModelList-method}
\alias{get_region}
\alias{get_region,PersephoneModel-method}
\alias{get_region,PersephoneModelList-method}
\alias{get_label}
\alias{get_label,PersephoneModel-method}
\alias{get_label,PersephoneModelList-method}
\alias{get_formula}
\alias{get_formula,PersephoneModel-method}
\alias{get_formula,PersephoneModelList-method}
\alias{get_model}
\alias{get_model,PersephoneModel-method}
\alias{get_model,PersephoneModelList-method}
\alias{get_crops}
\alias{get_crops,PersephoneModel-method}
\alias{get_crops,PersephoneModelList-method}
\alias{get_index}
\alias{get_index,PersephoneModelList-method}
\alias{get_stages}
\alias{get_stages,PersephoneModel-method}
\alias{get_stages,PersephoneModelList-method}
\alias{get_seasons}
\alias{get_seasons,PersephoneModel-method}
\alias{get_seasons,PersephoneModelList-method}
\title{Get object slots}
\usage{
get_class(object, ...)

\S4method{get_class}{PersephoneModel}(object, ...)

\S4method{get_class}{PersephoneModelList}(object, ...)

get_region(object, ...)

\S4method{get_region}{PersephoneModel}(object, ...)

\S4method{get_region}{PersephoneModelList}(object, ...)

get_label(object, ...)

\S4method{get_label}{PersephoneModel}(object, ...)

\S4method{get_label}{PersephoneModelList}(object, ...)

get_formula(object, ...)

\S4method{get_formula}{PersephoneModel}(object, ...)

\S4method{get_formula}{PersephoneModelList}(object, ...)

get_model(object, ...)

\S4method{get_model}{PersephoneModel}(object, ...)

\S4method{get_model}{PersephoneModelList}(object, ...)

get_crops(object, ...)

\S4method{get_crops}{PersephoneModel}(object)

\S4method{get_crops}{PersephoneModelList}(object)

get_index(object, ...)

\S4method{get_index}{PersephoneModelList}(object, crop = NULL)

get_stages(object, ...)

\S4method{get_stages}{PersephoneModel}(object)

\S4method{get_stages}{PersephoneModelList}(object)

get_seasons(object, ...)

\S4method{get_seasons}{PersephoneModel}(object)

\S4method{get_seasons}{PersephoneModelList}(object)
}
\arguments{
\item{object}{an object of class \code{PersephoneModel} or \code{PersephoneModelList}.}

\item{...}{extra arguments}

\item{crop}{character. A crop of choice.}
}
\value{
character. The \code{class()} of the \code{object}.
}
\description{
Get object slots
}
\section{Functions}{
\itemize{
\item \code{get_region()}: returns a character. The \code{region} slot of the \code{object}.

\item \code{get_label()}: returns a character. The \code{label} slot of the \code{object}.

\item \code{get_formula()}: returns a character. The \code{formula} slot of the \code{object}.

\item \code{get_model()}: returns a character. The \code{model} slot of the \code{object}.

\item \code{get_crops()}: returns a character. The \code{crop} slot of the \code{object}.

\item \code{get_index()}: returns a numeric The indices that correspond to the
\code{crop} slot of the \code{object}.

\item \code{get_stages()}: returns a factor. The \code{stages} included in \code{data}.

\item \code{get_seasons()}: returns a numeric. The \code{seasons} included in \code{data}.

}}
\examples{
\dontrun{
# Create a Region object
library(cronus)
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))

# Create a model
object <- new("PersephoneBin",
              label = "Base_logit",
              region = region,
              crop = "Corn",
              data = progress_ne$Corn,
              link = "logit",
              formula = "CumPercentage ~ Time + agdd + adayl") # PersephoneModel

# Fit
object <- fit(object)

# Get slots
get_class(object)
get_label(object)
get_formula(object)
get_model(object)
get_crops(object)
get_region(object)
get_index(object, crop = "Corn")
get_stages(object)
get_seasons(object)
}
}
