% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{get_combinations}
\alias{get_combinations}
\title{Get Combinations}
\usage{
get_combinations(x, prop, maxsam = Inf, seed = 1)
}
\arguments{
\item{x}{a vector with the combination elements.}

\item{prop}{numeric. the proportion of elements to be drawn.}

\item{maxsam}{numeric. maximum sample of combinations}

\item{seed}{numeric. \code{set.seed()} argument.}
}
\value{
matrix. Each column of the matrix is a different combination.
}
\description{
Get Combinations
}
\examples{
\dontrun{
get_combinations(1:10, 0.25, maxsam = 7, seed = 2938)
}
}
