% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_crops}
\alias{get_crops}
\alias{get_crops,PersephoneModel-method}
\alias{get_crops,PersephoneModelList-method}
\alias{get_region}
\alias{get_region,PersephoneModel-method}
\alias{get_region,PersephoneModelList-method}
\alias{get_index}
\alias{get_index,PersephoneModelList-method}
\alias{get_stages}
\alias{get_stages,PersephoneModel-method}
\alias{get_stages,PersephoneModelList-method}
\alias{get_seasons}
\alias{get_seasons,PersephoneModel-method}
\alias{get_seasons,PersephoneModelList-method}
\title{Tools}
\usage{
get_crops(object, ...)

\S4method{get_crops}{PersephoneModel}(object)

\S4method{get_crops}{PersephoneModelList}(object)

get_region(object, ...)

\S4method{get_region}{PersephoneModel}(object)

\S4method{get_region}{PersephoneModelList}(object)

get_index(object, ...)

\S4method{get_index}{PersephoneModelList}(object, crop = NULL)

get_stages(object, ...)

\S4method{get_stages}{PersephoneModel}(object)

\S4method{get_stages}{PersephoneModelList}(object)

get_seasons(object, ...)

\S4method{get_seasons}{PersephoneModel}(object)

\S4method{get_seasons}{PersephoneModelList}(object)
}
\arguments{
\item{object}{S4 object. A PersephoneModel or PersephoneModelList.}

\item{...}{extra arguments}

\item{crop}{character. A crop of choice.}
}
\value{
A vector or list of vectors
}
\description{
Access PersephoneModel slots and information.
}
\examples{
\dontrun{
get_crops(object)
get_region(object)
get_index(object, crop = "Corn")
get_stages(object)
get_seasons(object)
}
}
