% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_hts.R
\name{hierarchicalTimeSeries}
\alias{hierarchicalTimeSeries}
\title{R6 Class for hierarchical time series}
\description{
Combine mutliple objects of persephone objects into a new
persephone object. The resulting time series can perform direct and
indirect adjustments
}
\examples{
\dontrun{
obj_x13 <- per_x13(AirPassengers, "RSA3")

ht <- per_hts(a = obj_x13, b = obj_x13, method = "x13")
ht$run()
ht$adjusted_direct
ht$adjusted_indirect

ht2 <- per_hts(a = ht, b = obj_x13)
ht2$run()
ht2$adjusted_direct
ht2$adjusted_indirect
}

## ------------------------------------------------
## Method `hierarchicalTimeSeries$run`
## ------------------------------------------------

per_x13(AirPassengers)$run()
}
\section{Super class}{
\code{\link[persephone:persephone]{persephone::persephone}} -> \code{hierarchicalTimeSeries}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{components}}{the sub series of the hierarchical time series}

\item{\code{weights}}{the weights used for aggregating components}

\item{\code{indirect}}{wether to use direct or indirect adjustement}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{adjusted_indirect}}{results from the indirect adjustment where
all components are adjusted and then aggregated}

\item{\code{adjusted}}{results from the seasonal adjustment}

\item{\code{forecasts}}{get forecasts}

\item{\code{forecasts_indirect}}{get forecasts according to indirect adjustments}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{hierarchicalTimeSeries$new()}}
\item \href{#method-run}{\code{hierarchicalTimeSeries$run()}}
\item \href{#method-print}{\code{hierarchicalTimeSeries$print()}}
\item \href{#method-set_options}{\code{hierarchicalTimeSeries$set_options()}}
\item \href{#method-iterate}{\code{hierarchicalTimeSeries$iterate()}}
\item \href{#method-get_component}{\code{hierarchicalTimeSeries$get_component()}}
\item \href{#method-generate_qr_table}{\code{hierarchicalTimeSeries$generate_qr_table()}}
\item \href{#method-clone}{\code{hierarchicalTimeSeries$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="plot">}\href{../../persephone/html/persephone.html#method-plot}{\code{persephone::persephone$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="plotResiduals">}\href{../../persephone/html/persephone.html#method-plotResiduals}{\code{persephone::persephone$plotResiduals()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="plotSeasIrrCal">}\href{../../persephone/html/persephone.html#method-plotSeasIrrCal}{\code{persephone::persephone$plotSeasIrrCal()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="updateParams">}\href{../../persephone/html/persephone.html#method-updateParams}{\code{persephone::persephone$updateParams()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
create a new hierarchical time series
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$new(
  ...,
  method = c("tramoseats", "x13"),
  userdefined = NULL,
  spec = NULL,
  list = NULL,
  weights = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{one or more objects which are either of class persephone or
can be coerced to persephone objects with as_persephone. If more than
one element is supplied, the underlying time series must have the same
time instances. All elements supplied in ... must be named.}

\item{\code{method}}{specifies the method to be used for the direct adjustment
of the aggregate series. tramoseats or x13}

\item{\code{userdefined}}{passed as the userdefined argument to tramoseats() or
x13()}

\item{\code{spec}}{a model specification returned by x13_spec() or
tramoseats_spec()}

\item{\code{list}}{a list of persephone objects as alternative input to \code{...}.
This argument can also handle mts objects}

\item{\code{weights}}{either a vector if the same weight is used for all time
points or a list of ts objects or a mts object if the weight varies
for different time points. They must have the same length as the
number of components.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\subsection{Method \code{run()}}{
run the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$run(verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{if \code{FALSE} (the default), the results of the run will
be returned invisibly}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{per_x13(AirPassengers)$run()
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
print a hierarchical timeseries to screen
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_options"></a>}}
\subsection{Method \code{set_options()}}{
sets options for all entries of the dependency tree
recursively (if recursive = TRUE). See
vignette("persephone-hierarchical") for more details.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$set_options(
  userdefined = NA,
  spec = NA,
  recursive = TRUE,
  component = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userdefined}}{additional outputs to generate while running. See
\code{\link[=x13]{x13()}} and \code{\link[=tramoseats]{tramoseats()}}.}

\item{\code{spec}}{specifications generated by \code{x13_spec()} or
\code{tramoseats_spec()}}

\item{\code{recursive}}{apply this setting to all subseries as well?}

\item{\code{component}}{which component to modify.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iterate"></a>}}
\subsection{Method \code{iterate()}}{
iterate over all components
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$iterate(
  fun,
  as_table = FALSE,
  component = "",
  unnest = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{a function that takes a persephone object as a parameter}

\item{\code{as_table}}{if true, the return value of this method will be coerced
to a data.frame}

\item{\code{component}}{the id of the component}

\item{\code{unnest}}{if \code{as_table = FALSE}, converts the return value from a
nested list into a flat list}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
this functin is similar to \code{lapply()} in the sense that it
can be used to apply a function to several persephone objects
simultaniusely
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_component"></a>}}
\subsection{Method \code{get_component()}}{
extract a component series
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$get_component(component_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{component_id}}{the id of a component}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_qr_table"></a>}}
\subsection{Method \code{generate_qr_table()}}{
Generate a table for the eurostat quality report
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$generate_qr_table(component = "")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{component}}{(optional) a sub-component to create the report for}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
