% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{lesions_to_percents}
\alias{lesions_to_percents}
\title{Calculate lesion prevalence by age interval}
\usage{
lesions_to_percents(cemetery_df, group_vars)
}
\arguments{
\item{cemetery_df}{Data frame of cemetery data with age and lesion columns}

\item{group_vars}{Character vector of grouping variable names}
}
\value{
Data frame with Lesion_Percent by Age_Interval and groups
}
\description{
Calculate lesion prevalence by age interval
}
