% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_batch.R
\name{per_batch}
\alias{per_batch}
\title{Define a multiple time series}
\usage{
per_batch(
  ...,
  method = c("tramoseats", "x13"),
  userdefined = NULL,
  spec = NULL,
  list = NULL
)
}
\arguments{
\item{...}{ne or more objects which are either of class persephone or can be
coerced to persephone objects with as_persephone. If more than one element
is supplied, the underlying time series must have the same time instances.
All elements supplied in ... must be named.}

\item{method}{specifies the method to be used for the adjustment of
the aggregate series. tramoseats or x13}

\item{userdefined}{passed as the userdefined argument to tramoseats() or x13()}

\item{spec}{a model specification returned by x13_spec() or
tramoseats_spec()}

\item{list}{a list of persephone objects as alternative input to \code{...}. This
argument can also handle mts objects}
}
\description{
Combine mutliple objects of persephone objects into a new persephone batch object.
On the resulting objects, you can perform seasonal adjustment.
}
