% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_hts.R
\name{per_hts}
\alias{per_hts}
\title{Define a hierarchical time series}
\usage{
per_hts(
  ...,
  method = c("tramoseats", "x13"),
  userdefined = NULL,
  spec = NULL,
  list = NULL,
  weights = NULL
)
}
\arguments{
\item{...}{ne or more objects which are either of class persephone or can be
coerced to persephone objects with as_persephone. If more than one element
is supplied, the underlying time series must have the same time instances.
All elements supplied in ... must be named.}

\item{method}{specifies the method to be used for the direct adjustment of
the aggregate series. tramoseats or x13}

\item{userdefined}{passed as the userdefined argument to tramoseats() or x13()}

\item{spec}{a model specification returned by x13_spec() or
tramoseats_spec()}

\item{list}{a list of persephone objects as alternative input to \code{...}. This
argument can also handle mts objects}

\item{weights}{either a vector if the same weight is used for all time
points or a list of ts objects or a mts object if the weight varies for
different time points. They must have the same length as the number of
components.}
}
\description{
Combine mutliple objects of persephone objects into a new persephone object.
The resulting time series can perform direct and indirect adjustments.
}
