% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_tramo.R
\name{per_tramo}
\alias{per_tramo}
\title{tramoseats class}
\usage{
per_tramo(
  ts,
  template = c("RSAfull", "RSA0", "RSA1", "RSA2", "RSA3", "RSA4", "RSA5"),
  userdefined = NULL,
  ...
)
}
\arguments{
\item{ts}{a time series}

\item{template}{passed as the \code{spec} argument to \code{\link[=tramoseats_spec]{tramoseats_spec()}}}

\item{userdefined}{passed as the \code{userdefined} argument to \code{\link[=tramoseats]{tramoseats()}}}

\item{...}{passed to \code{\link[=tramoseats_spec]{tramoseats_spec()}}}
}
\description{
an R6 class that represents a tramoseats model
}
\section{Inherits}{
 \link{persephone}
}

\examples{
data(AirPassengers, package = "datasets")

obj <- per_tramo(AirPassengers, "RSA3")
obj$run()
obj$ts
}
