% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_x13.R
\name{per_x13}
\alias{per_x13}
\title{x13 class}
\usage{
per_x13(
  ts,
  template = c("RSA5c", "RSA0", "RSA1", "RSA2c", "RSA3", "RSA4c", "X11"),
  userdefined = NULL,
  ...
)
}
\arguments{
\item{ts}{a time series}

\item{template}{passed as the \code{spec} argument to \code{\link[=x13_spec]{x13_spec()}}}

\item{userdefined}{passed as the \code{userdefined} argument to \code{\link[=x13]{x13()}}}

\item{...}{passed to \code{\link[=x13_spec]{x13_spec()}}}
}
\description{
An R6 class that represents a x13 model.
}
\section{Inherits}{
 \link{persephone}
}

\examples{
data(AirPassengers, package = "datasets")

obj <- per_x13(AirPassengers, "RSA3")
obj$run()
obj$ts
}
