% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perBatch.R
\name{perBatch}
\alias{perBatch}
\title{Define a multiple time series}
\usage{
perBatch(..., list = NULL)
}
\arguments{
\item{...}{ne or more objects which are either of class persephone or can be
coerced to persephone objects with asPersephone. If more than one element
is supplied, the underlying time series must have the same time instances.
All elements supplied in ... must be named.}

\item{list}{a list of persephone objects as alternative input to \code{...}. This
argument can also handle mts objects}
}
\description{
Combine mutliple objects of persephone objects into a new persephone batch object.
On the resulting objects, you can perform seasonal adjustment.
}
