% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persephone-package.R
\docType{package}
\name{persephone-package}
\alias{persephone}
\alias{persephone-package}
\title{persephone: Agent-Based Model for Simulating Skeletal Lesion Formation and Mortality in Archaeological Populations}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Amy Anderson \email{first.last@example.com}

}
