% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persephone.R
\name{persephone}
\alias{persephone}
\title{persephone class}
\description{
Common interface for x13 and tramoseats
}
\section{Methods}{

\itemize{
\item \code{run()}: Runs the model
\item \code{updateParams(...)}: Updates the current parameters
}
}

\section{Read-Only fields}{

\itemize{
\item params
\item ts
\item output
}
}

