% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{PersephoneModel-class}
\alias{PersephoneModel-class}
\alias{PersephoneBinomial-class}
\alias{PersephoneCumLink-class}
\title{Persephone Model Classes}
\value{
An S4 object of the appropriate class.
}
\description{
Persephone Model Classes
}
\section{Slots}{

\describe{
\item{\code{region}}{character. A region of interest.}

\item{\code{data}}{list.}

\item{\code{model}}{list.}

\item{\code{fitted}}{list.}

\item{\code{formula}}{character.}

\item{\code{correlation}}{ANY.}

\item{\code{weights}}{ANY.}

\item{\code{link}}{character.}

\item{\code{nominal}}{character.}
}}

\examples{
\dontrun{
object <- new("PersephoneCumLink")
}
}
