% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perSingle.R
\name{persephoneSingle}
\alias{persephoneSingle}
\title{Class for single time series}
\description{
Class for single time series

Class for single time series
}
\details{
A common class for the wrapper classes \link{x13Single} and \link{tramoseatsSingle}
}
\section{Super class}{
\code{\link[persephone:persephone]{persephone::persephone}} -> \code{persephoneSingle}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{persephoneSingle$new()}}
\item \href{#method-run}{\code{persephoneSingle$run()}}
\item \href{#method-clone}{\code{persephoneSingle$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="fixModel">}\href{../../persephone/html/persephone.html#method-fixModel}{\code{persephone::persephone$fixModel()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="fixOutlier">}\href{../../persephone/html/persephone.html#method-fixOutlier}{\code{persephone::persephone$fixOutlier()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="generateQrTable">}\href{../../persephone/html/persephone.html#method-generateQrTable}{\code{persephone::persephone$generateQrTable()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="iterate">}\href{../../persephone/html/persephone.html#method-iterate}{\code{persephone::persephone$iterate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="plot">}\href{../../persephone/html/persephone.html#method-plot}{\code{persephone::persephone$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="plotResiduals">}\href{../../persephone/html/persephone.html#method-plotResiduals}{\code{persephone::persephone$plotResiduals()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="plotSeasIrrCal">}\href{../../persephone/html/persephone.html#method-plotSeasIrrCal}{\code{persephone::persephone$plotSeasIrrCal()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="print">}\href{../../persephone/html/persephone.html#method-print}{\code{persephone::persephone$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="setOptions">}\href{../../persephone/html/persephone.html#method-setOptions}{\code{persephone::persephone$setOptions()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="updateParams">}\href{../../persephone/html/persephone.html#method-updateParams}{\code{persephone::persephone$updateParams()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
create a new single object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephoneSingle$new(ts, template = "RSA0", userdefined = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ts}}{a time series object}

\item{\code{template}}{passed as the \code{spec} argument of \code{\link[=x13]{x13()}} or
\code{\link[=tramoseats]{tramoseats()}}}

\item{\code{userdefined}}{passed as the \code{userdefined} argument of \code{\link[=x13]{x13()}} or
\code{\link[=tramoseats]{tramoseats()}}}

\item{\code{...}}{passed to \code{\link[=x13_spec]{x13_spec()}} or \code{\link[=tramoseats_spec]{tramoseats_spec()}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\if{latex}{\out{\hypertarget{method-run}{}}}
\subsection{Method \code{run()}}{
run the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephoneSingle$run(verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{wether to display outputs after the run}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephoneSingle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
