% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perTramo.R
\name{perTramo}
\alias{perTramo}
\title{tramoseats class}
\usage{
perTramo(
  ts,
  template = c("rsafull", "rsa0", "rsa1", "rsa2", "rsa3", "rsa4", "rsa5"),
  userdefined = NULL,
  ...
)
}
\arguments{
\item{ts}{a time series}

\item{template}{passed as the \code{spec} argument to \code{\link[=tramoseats_spec]{tramoseats_spec()}}}

\item{userdefined}{passed as the \code{userdefined} argument to \code{\link[=tramoseats_fast]{tramoseats_fast()}}}

\item{...}{passed to \code{\link[=tramoseats_spec]{tramoseats_spec()}}}
}
\description{
an R6 class that represents a tramoseats model
}
\section{Inherits}{
 \link{persephone}
}

\examples{
data(AirPassengers, package = "datasets")

obj <- perTramo(AirPassengers, "rsa3")
obj$run()
obj$ts
}
