% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perX13.R
\name{perX13}
\alias{perX13}
\title{x13 class}
\usage{
perX13(
  ts,
  template = c("rsa4", "rsa0", "rsa1", "rsa2c", "rsa3", "rsa5c"),
  userdefined = NULL,
  ...
)
}
\arguments{
\item{ts}{a time series}

\item{template}{passed as the \code{name} argument to \code{\link[=x13_spec]{x13_spec()}}}

\item{userdefined}{passed as the \code{userdefined} argument to \code{\link[=x13_fast]{x13_fast()}}}

\item{...}{passed to \code{\link[=x13_spec]{x13_spec()}}}
}
\description{
An R6 class that represents a x13 model.
}
\section{Inherits}{
 \link{persephone}
}

\examples{
data(AirPassengers, package = "datasets")

obj <- perX13(AirPassengers, "rsa3")
obj$run()
obj$ts
}
