% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pi_sa}
\alias{pi_sa}
\title{Production in industry - Seasonally and calendar adjusted}
\format{A multivariate time series objects with 30 time series of length 234.}
\usage{
data(pi_sa)
}
\description{
A multivariate time series object containing the volume index of production (Index, 2015=100)
for mining and quarrying; manufacturing; electricity, gas, steam and air conditioning supply
for the EU-28 countries and EU-28 and EA-19 (Euro Area) aggregates.
}
\references{
Extracted on	03.09.19 from \url{https://ec.europa.eu/eurostat/web/short-term-business-statistics/data/database}
}
\keyword{datasets}
