% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pi_unadj}
\alias{pi_unadj}
\title{Production in industry - Unadjusted data (i.e. neither seasonally adjusted
nor calendar adjusted data)}
\format{A multivariate time series objects with 30 time series of length 234.}
\usage{
data(pi_unadj)
}
\description{
A multivariate time series object containing the volume index of production
(Index, 2015=100) for mining and quarrying; manufacturing; electricity, gas,
steam and air conditioning supply for the EU-28 countries and EU-28 and EA-19
(Euro Area) aggregates.
}
\references{
Extracted on	03.09.19 from \url{https://ec.europa.eu/eurostat/web/short-term-business-statistics/data/database}
}
\keyword{datasets}
