% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_contrast}
\alias{plot_contrast}
\alias{plot_contrast,data.frame,ProgressBM-method}
\title{Plot model contrast}
\usage{
plot_contrast(x, y, ...)

\S4method{plot_contrast}{data.frame,ProgressBM}(
  x,
  y,
  save = FALSE,
  path = cronus::get_path_hermes(),
  file = "plot.pdf",
  width = 15,
  height = 8
)
}
\arguments{
\item{x}{data.frame.}

\item{y}{an object of class \code{ProgressModel}.}

\item{...}{extra arguments}

\item{save}{logical. Should the plot be saved?}

\item{path}{character. The directory in which the plot will be saved.}

\item{file}{character. The file name.}

\item{width}{numeric. The width of the plot in inches.}

\item{height}{numeric. The height of the plot in inches.}
}
\value{
A plot, created with \code{ggplot2} The plot is returned wrapped inside
\code{invisible()}.
}
\description{
Plot model contrast
}
\examples{
\dontrun{
# Create a Region object
library(cronus)
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))

# Create a model
object1 <- new("ProgressBM",
               region = region,
               crop = "Corn",
               data = data_progress$Corn,
               formula = "CumPercentage ~ Time + agdd") # ProgressModel

# Create another model
object2 <- new("ProgressCLM",
               region = region,
               crop = "Soybeans",
               data = data_progress$Soybeans,
               formula = "Stage ~ Time + agdd + adayl") # ProgressModel

# Concatenate the models
object <- c(object1, object2) # ProgressModelList

# Fit
object <- fit(object)

# Plot
plot(object, cumulative = TRUE, seasons = 2002)

# Predict
predict(object, data_progress)

# Evaluate
object <- evaluate(object, maxsam = 100, seed = 1)
plot_metrics(object)

# Summarize
summary(object)

# Report
report(object, name = "example_report", path = getwd())
}
}
