% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_IRS.R
\name{plot_IRS}
\alias{plot_IRS}
\title{A function to save 4 core Impact response surfaces for paper,
given a specific functional form for a specific ensemble
and the corresponding coefficients for mu and sigma.}
\usage{
plot_IRS(coefficients, SDcoefficients, functionalform, sdfunctionalform,
  baseCO2, baseTemp, basePrecip, name, mesh = 101, figpath = "figures/irs",
  legendsave = FALSE, high = FALSE, low = FALSE)
}
\arguments{
\item{coefficients}{The point values of coefficients for mu}

\item{SDcoefficients}{The point values of coefficients for sigma}

\item{functionalform}{The name functional form for the mean of the likelihood}

\item{sdfunctionalform}{The name of the functional form for the standard deviation of the likelihood}

\item{baseCO2}{The baseline value of CO2 concentration}

\item{baseTemp}{The baseline value of temperature changes}

\item{basePrecip}{The baseline value of precipitation changes}

\item{name}{The name of the ensemble under consideration}

\item{mesh}{Define how many steps in each direction to make the CTW grid for IRS plotting, defaults to 101}

\item{figpath}{The location to write the IRSplots, defaults to "figures/irs"}

\item{legendsave}{Whether to save the master legend of IRS changes, defaults to FALSE}

\item{high}{If you want to return the high response case instead of the mean, defaults to FALSE}

\item{low}{If you want to return the low response case instead of the mean, defaults to FALSE}
}
\description{
A function to save 4 core Impact response surfaces for paper,
given a specific functional form for a specific ensemble
and the corresponding coefficients for mu and sigma.
}
