% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map_emulated_yields.R
\name{plot_map_emulated_yields}
\alias{plot_map_emulated_yields}
\title{A function to save 4 core Impact response surfaces for paper,
given a specific functional form for a specific ensemble
and the corresponding coefficients for mu and sigma.}
\usage{
plot_map_emulated_yields(crop, coefpath = "data/emulatorcoefficients",
  ctwpath = "data/climatedata/hadgem2es", mapsavepath = "figures/maps",
  plotyear = 2085, plotirr = "RFD", func, sdfunc, high = FALSE,
  low = FALSE)
}
\arguments{
\item{crop}{The crop to calculate changes for, including looking up coefficients}

\item{coefpath}{The location of the coefficients, defaults to "data/emulatorcoefficients"}

\item{ctwpath}{The location of the driving CTW data, defaults to "data/climatedata/hadgem2es"}

\item{mapsavepath}{The location to save the output maps, defaults to "figures/maps"}

\item{plotyear}{The year to plot, defaults to 2085}

\item{plotirr}{The type of crop to plot, RFD or IRR. Defaults to RFD}

\item{func}{The name of the functional form for mu, for looking up coefficients}

\item{sdfunc}{The name of the functional form for sigma, for looking up coefficients}

\item{high}{If you want to return the high response case instead of the mean, defaults to FALSE}

\item{low}{If you want to return the low response case instead of the mean, defaults to FALSE}
}
\description{
A function to save 4 core Impact response surfaces for paper,
given a specific functional form for a specific ensemble
and the corresponding coefficients for mu and sigma.
}
