% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_metrics}
\alias{plot_metrics}
\alias{plot_metrics,PersephoneModel-method}
\alias{plot_metrics,PersephoneModelList-method}
\title{Plot Metrics}
\usage{
plot_metrics(object, ...)

\S4method{plot_metrics}{PersephoneModel}(
  object,
  title = NULL,
  save = FALSE,
  dir = getwd(),
  file = "plot.pdf",
  width = 15,
  height = 8,
  print_plots = TRUE
)

\S4method{plot_metrics}{PersephoneModelList}(
  object,
  ncol = NULL,
  nrow = NULL,
  save = FALSE,
  dir = getwd(),
  file = "plot.pdf",
  width = 15,
  height = 8,
  ...
)
}
\arguments{
\item{object}{an S4 object. The model of interest.}

\item{...}{extra arguments.}

\item{save}{logical. Should the plot be saved?}

\item{dir}{character. The directory in which the plot will be saved.}

\item{file}{character. The file name.}

\item{width}{numeric. The width of the plot in inches.}

\item{height}{numeric. The height of the plot in inches.}

\item{print_plots}{logical. Should the plots be printed?}

\item{ncol}{numeric. Number of rows in the plot layout.}

\item{nrow}{numeric. Number of columns in the plot layout.}
}
\value{
A plot, created with \code{ggplot2} and \code{ggpubr}. The plot is
returned wrapped inside \code{invisible()}.
}
\description{
Create plots of the model metrics.
}
