% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_survival_curves}
\alias{plot_survival_curves}
\title{Plot survival curves by lesion status}
\usage{
plot_survival_curves(survival_data)
}
\arguments{
\item{survival_data}{Data frame from run_survival_analysis}
}
\value{
A ggplot object
}
\description{
Plot survival curves by lesion status
}
