% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSeasIrrCal.R
\name{plotSeasIrrCal}
\alias{plotSeasIrrCal}
\title{irregular component and calendar effects for a persephone object}
\usage{
plotSeasIrrCal(x)
}
\arguments{
\item{x}{an object of class \link{persephone}.}
}
\description{
The function generates an interactive time series plot of the seasonal
component, irregular component and calendar effects
for a \link{persephone} object together with one year forecasts
}
