% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSIC.R
\name{plotSIC.persephone}
\alias{plotSIC.persephone}
\title{Interactive plot of the seasonal component, irregular component and calendar effects for a persephone object}
\usage{
plotSIC.persephone(object, main = NULL, forecasts = TRUE,
  rangeSelector = TRUE, drawPoints = FALSE, annualComparison = NULL)
}
\arguments{
\item{object}{an object of class \code{\link{persephone}}.}

\item{main}{plot title}

\item{forecasts}{logical flag indicating if forecasts should be plotted}

\item{rangeSelector}{logical flag specifying if a range selector should be included in the plot}

\item{drawPoints}{logical flag indicating if a small dot should be drawn at each point, in addition to a line going through the point.}

\item{annualComparison}{integer corresponding to the month or quarter which should be highlighted in the plot for every year.}
}
\value{
Returns an object of class \code{dygraphs}.
}
\description{
Produces a dygraph (see the \href{https://rstudio.github.io/dygraphs/}{online documentation}
for more detail) for objects of class \code{persephone}.
The function generates an interactive time series plot of the seasonal component, irregular component and calendar effects
for a persephone object together with one year forecasts
}
\examples{
data(myseries, package = "RJDemetra")
# Generate a persephone object, in this case ??? an x13Single object
obj <- x13Single$new(myseries, "RSA1", userdefined=c("y","t","sa",
                                                      "s","i","cal",
                                                      "y_f","t_f","sa_f",
                                                      "s_f","i_f","cal_f",
                                                      "preprocessing.model.y_f",
                                                      "preprocessing.model.y_ef")) ## noch zu x13Single dazugeben
obj$run()
# Plot after run
obj$SICplot()

}
