% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSIC.R
\name{plotSIC}
\alias{plotSIC}
\title{Interactive plot of the seasonal component, irregular component and calendar effects for a persephone object}
\usage{
plotSIC(object, main = NULL, forecasts = TRUE, rangeSelector = TRUE,
  drawPoints = FALSE, annualComparison = NULL)
}
\arguments{
\item{object}{an object of class \link{persephone}.}

\item{main}{plot title}

\item{forecasts}{logical flag indicating if forecasts should be plotted}

\item{rangeSelector}{logical flag specifying if a range selector should be included in the plot}

\item{drawPoints}{logical flag indicating if a small dot should be drawn at each point, in addition to a line going through the point.}

\item{annualComparison}{integer corresponding to the month or quarter which should be highlighted in the plot for every year.}
}
\value{
Returns an object of class \code{dygraphs}.
}
\description{
Produces a dygraph (see the \href{https://rstudio.github.io/dygraphs/}{online documentation}
for more detail) for objects of class \link{persephone}.
The function generates an interactive time series plot of the seasonal component, irregular component and calendar effects
for a persephone object together with one year forecasts
}
\examples{
data(AirPassengers, package = "datasets")
# Generate a persephone object, in this case an x13Single object
obj <- x13Single$new(AirPassengers, "RSA1")

obj$run()
# Plot after run
plotSIC(obj)

}
