% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSiRatios.R
\name{plotSiRatios}
\alias{plotSiRatios}
\title{Plot of SI-ratios by period}
\usage{
plotSiRatios(x)
}
\arguments{
\item{x}{An object of class \code{persephone}}
}
\value{
An object of class \code{highchart}
}
\description{
Generate a \code{hchart} showing the si ratios and seasonal factors.
}
\examples{
x <- perX13(AirPassengers)
x$run()
plotSiRatios(x)
}
