% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSiRatios_old.R
\name{plotSiRatios_old}
\alias{plotSiRatios_old}
\title{Interactive plot of SI-ratios (and, in case of x11, of seasonal factors) by
period}
\usage{
plotSiRatios_old(x, main = NULL, interactive = TRUE, ...)
}
\arguments{
\item{x}{an object of class \link{persephone}.}

\item{main}{plot title}

\item{interactive}{If the return value would be a \code{ggplot} object, wrap it in
\link[plotly:ggplotly]{plotly::ggplotly} before returning.}

\item{...}{other plotting parameters to affect the plot. Not currently used.}
}
\value{
Returns an object of class \code{ggplot} or \code{plotly}
}
\description{
Produces a \code{ggplot}/\code{plotly} object for objects of class \link{persephone}.
If the \code{persephone} object is of class \code{tramoseatsSingle}, the function
generates an interactive plot of the SI-ratios and their mean. Otherwise,
i.e. if the object is of class \code{x13Single}, the SI-ratios, the replaced
SI-ratios, the seasonal factors and the mean of the seasonal factors are
plotted.
}
\examples{

data(AirPassengers, package = "datasets")
# Generate a persephone object, in this case an x13Single object
obj <- perX13(AirPassengers, "rsa1")
obj$run()
# Plot the SI-ratios after run
plotSiRatios(obj)

data(UKgas, package = "datasets")
# Generate a persephone object, in this case a tramoseatsSingle object
obj2 <- perTramo(UKgas, "rsa3")
obj2$run()
plotSiRatios(obj2)

}
