% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpectrum.R
\name{plotSpectrum}
\alias{plotSpectrum}
\title{Spectral Plots}
\usage{
plotSpectrum(
  x,
  tsType = c("original", "sa", "irregular", "residuals"),
  plotType = c("arSpec", "arSpecBars", "periodogram"),
  maxobs = NULL,
  main = NULL,
  n.freq = NULL,
  order = 30
)
}
\arguments{
\item{x}{an object of class \link{persephone}.}

\item{tsType}{character (\code{"original"},\code{"sa"},\code{"irregular"},\code{"residuals"})
selecting the preferred type of time series to plot. A numeric
value (\code{1:4}) corresponding to one of these characters is also accepted.}

\item{plotType}{character (\code{"arSpec"},\code{"arSpecBars"},\code{"periodogram"})
selecting the preferred type of plot, see Details. A numeric
value (\code{1:3}) corresponding to one of these characters is also accepted.}

\item{maxobs}{maximum number of observations of the time series. If \code{NULL},
\code{maxobs} equals the length of the time series for plotTypes
\code{"arSpec"} and \code{"periodogram"} and \code{maxobs=96} for plotType \code{"arSpecBars"}.}

\item{main}{plot title}

\item{n.freq}{the number of frequencies, i.e. the number of points at which
to plot. If \code{NULL}, \code{n.freq=301} for plotTypes \code{"arSpec"} and \code{"periodogram"}
and \code{n.freq=61} for plotType \code{"arSpecBars"}.}

\item{order}{order of the AR model.}
}
\description{
Spectral Plots
}
\examples{
x <- perX13(AirPassengers)
x$run()
plotSpectrum(x)
plotSpectrum(x, plotType = "periodogram")
plotSpectrum(x, tsType = "irregular")
}
