% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_classes.R
\docType{class}
\name{ProgressModel-class}
\alias{ProgressModel-class}
\alias{ProgressBM-class}
\alias{ProgressBMM-class}
\alias{ProgressCLM-class}
\alias{ProgressCLMM-class}
\alias{ProgressSRF-class}
\alias{ProgressMRF-class}
\title{\code{persephone} Model classes}
\value{
An S4 object of the appropriate class.
}
\description{
Package \code{persephone} defines a number of model classes. The main functions
of the package receive a model object and edit its slots.
}
\details{
\itemize{
\item ProgressBM is based on \code{glm()}.
\item ProgressBMM is based on \code{lme4::glmer()}.
\item ProgressCLM is based on \code{ordinal::clm()}.
\item ProgressCLMm is based on \code{ordinal::clm2()}.
}
}
\section{Slots}{

\describe{
\item{\code{label}}{character. The object's label.}

\item{\code{region}}{an object of class Region, defined in package \code{cronus}. A
region of interest.}

\item{\code{data}}{an object of class \code{Progress}, defined in package \code{cronus}. The
data on which the model will be fitted.}

\item{\code{model}}{ANY. An object holding information about the fitting. Filled by
function \code{fit()}.}

\item{\code{fitted}}{an object of class \code{Progress}, defined in package \code{cronus}.
The fitted data. Filled by function \code{fit()}.}

\item{\code{metrics}}{an object of class \code{ProgressMetrics}. The model performance
metrics. Filled by function \code{evaluate()}.}

\item{\code{formula}}{character. The model formula. See details.}

\item{\code{nominal}}{character. The nominal effects. See details.}

\item{\code{scale}}{character. The scale effects. See details.}

\item{\code{link}}{character. The link function used.}

\item{\code{threshold}}{character. The type of thresholds used. See details.}

\item{\code{nAGQ}}{numeric. The number of quadrature points to be used in the
adaptive Gauss-Hermite quadrature approximation to the marginal likelihood.
See details.}
}}

\examples{
\dontrun{
# Create a Region object
library(cronus)
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))

# Create a model
object1 <- new("ProgressBM",
               region = region,
               crop = "Corn",
               data = data_progress$Corn,
               formula = "CumPercentage ~ Time + agdd") # ProgressModel

# Create another model
object2 <- new("ProgressCLM",
               region = region,
               crop = "Soybeans",
               data = data_progress$Soybeans,
               formula = "Stage ~ Time + agdd + adayl") # ProgressModel

# Concatenate the models
object <- c(object1, object2) # ProgressModelList
}
}
\seealso{
\code{\link[=fit]{fit()}}, \code{\link[=predict]{predict()}},
\code{\link[=evaluate]{evaluate()}}, \code{\link[ordinal:clm]{ordinal::clm()}}, \code{\link[ordinal:clmOld]{ordinal::clm2()}}
}
