% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep.R
\name{read_model_sweep}
\alias{read_model_sweep}
\title{Read saved sweep results from disk}
\usage{
read_model_sweep(
  root_output_directory,
  target_param,
  target_param_values,
  data_file = "sim_cemetery.csv"
)
}
\arguments{
\item{root_output_directory}{Root directory containing sweep outputs}

\item{target_param}{Character name of the swept parameter}

\item{target_param_values}{Vector of parameter values to read}

\item{data_file}{Filename to read from each replicate directory}
}
\value{
Nested list of data frames
}
\description{
Read saved sweep results from disk
}
