% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R, R/helpers.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{hchart}
\alias{hc_title}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}

  \item{highcharter}{\code{\link[highcharter:highcharter-exports]{\%>\%}}, \code{\link[highcharter]{hc_title}}, \code{\link[highcharter]{hchart}}}
}}

