% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_add_columns.R
\name{repeat_add_columns}
\alias{repeat_add_columns}
\title{A function to repeat for added columns
Cribbed from gcamdata}
\usage{
repeat_add_columns(x, y)
}
\arguments{
\item{x}{dataframe of columns to be repeated for each added column of y}

\item{y}{dataframe of columns to add to x.}
}
\description{
A function to repeat for added columns
Cribbed from gcamdata
}
