% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\alias{report,PersephoneModel-method}
\alias{report,PersephoneModelList-method}
\title{Create reports}
\usage{
report(object, ...)

\S4method{report}{PersephoneModel}(object, name, path = getwd(), bell = 2)

\S4method{report}{PersephoneModelList}(object, name, path = getwd(), bell = 2)
}
\arguments{
\item{object}{S4 object. The model of interest.}

\item{...}{extra arguments.}

\item{name}{character. The html file name.}

\item{path}{character. The path to store the html report.}

\item{bell}{numeric. A sound notifying the user that the report(s) is (are) ready.}
}
\value{
nothing. The html report(s) is (are) created and stored.
}
\description{
Create reports
}
\examples{
\dontrun{
report(object, name = "myreport", path = getwd())
}
}
