% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rmspe}
\alias{rmspe}
\title{Root Mean Square Prediction Error}
\usage{
rmspe(x, y)
}
\arguments{
\item{x}{data.frame. A data.frame with columns named "Stage", "Season",
"Time", "Date" and "Percentage".}

\item{y}{data.frame. Same structure as \code{x}.}
}
\value{
A data.frame with columns Stage, Time and rmspe.
}
\description{
Root Mean Square Prediction Error
}
