% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep.R
\name{run_model_reps}
\alias{run_model_reps}
\title{Run the model multiple times with identical parameters}
\usage{
run_model_reps(params, root_output_directory, numreps = 10)
}
\arguments{
\item{params}{Named list of simulation parameters}

\item{root_output_directory}{Root directory for saving replicate outputs}

\item{numreps}{Number of replicates to run}
}
\description{
Run the model multiple times with identical parameters
}
