% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweep.R
\name{run_sweep}
\alias{run_sweep}
\title{Run a parameter sweep over lesion formation rates}
\usage{
run_sweep(mortality_regime, lesion_rates, reps, rmr)
}
\arguments{
\item{mortality_regime}{Data frame with Siler parameters}

\item{lesion_rates}{Vector of lesion formation rates to sweep}

\item{reps}{Number of replicates per rate}

\item{rmr}{Relative mortality risk value}
}
\value{
Combined data frame of all simulation results
}
\description{
Run a parameter sweep over lesion formation rates
}
