% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{scenario_lineplot}
\alias{scenario_lineplot}
\title{Line plot of average lesion prevalence by scenario}
\usage{
scenario_lineplot(plot_summary_data)
}
\arguments{
\item{plot_summary_data}{Summary data frame with mean outcomes}
}
\value{
A ggplot object
}
\description{
Line plot of average lesion prevalence by scenario
}
