% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary,PersephoneModelList-method}
\alias{summary,PersephoneBinomial-method}
\alias{summary,PersephoneCumLink-method}
\title{Persephone model summary}
\usage{
summary(object, ...)

\S4method{summary}{PersephoneModelList}(object, ...)

\S4method{summary}{PersephoneBinomial}(object, ...)

\S4method{summary}{PersephoneCumLink}(object, ...)
}
\arguments{
\item{object}{S4 object. The model of interest.}

\item{...}{extra arguments.}
}
\value{
S4 object. The model of interest.
}
\description{
Persephone model summary
}
\examples{
\dontrun{
region <- c(name = "nebraska", type = "us state")
formula <- "Percentage ~ Time"
object <- new("PersephoneBinomial", region = region, formula = formula)
object <- fit(object, data = data)
summary(object)
}
}
