% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{sweep_lineplot}
\alias{sweep_lineplot}
\title{Line plot of lesion prevalence across sweep replicates}
\usage{
sweep_lineplot(plot_data, plot_color)
}
\arguments{
\item{plot_data}{Data frame from lesions_to_percents}

\item{plot_color}{Color for the lines}
}
\value{
A ggplot object
}
\description{
Line plot of lesion prevalence across sweep replicates
}
