% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{weights_pi_ea19}
\alias{weights_pi_ea19}
\title{Production Weightings: Countries' share as a \% of EA-19 for Total Industry
(except construction)}
\format{A data.frame with 28 rows and 2 columns.}
\usage{
data(weights_pi_ea19)
}
\description{
A dataset containing the EA-19 (Euro Area) countries and the respective
weights
}
\details{
\itemize{
\item country. EU-28 country
\item weight. weight of the respective country
}
}
\references{
See \url{https://ec.europa.eu/eurostat/documents/2995521/9490436/4-14012019-AP-EN.pdf/9d606acd-97bc-4a5b-8c4c-a270888aab0a}
}
\keyword{datasets}
