% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{c,PersephoneModel-method}
\alias{c,PersephoneModel-method}
\title{Concatenate}
\usage{
\S4method{c}{PersephoneModel}(x, ...)
}
\arguments{
\item{x}{an object of class \code{PersephoneModel}.}

\item{...}{extra models to concatenate.}
}
\value{
an object of class \code{PersephoneModelList}.
}
\description{
Concatenate PersephoneModel objects into a PersephoneModelList.
}
\examples{
\dontrun{
# Create a Region object
library(cronus)
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))

# Create a model
object1 <- new("PersephoneBin",
               region = region,
               crop = "Corn",
               data = progress_ne$Corn,
               formula = "CumPercentage ~ Time + agdd") # PersephoneModel

# Create another model
object2 <- new("PersephoneCumLink",
               region = region,
               crop = "Soybeans",
               data = progress_ne$Soybeans,
               formula = "Stage ~ Time + agdd + adayl") # PersephoneModel

# Concatenate the models
object <- c(object1, object2) # PersephoneModelList

# Fit
object <- fit(object)

# Plot
plot(object, cumulative = TRUE, seasons = 2002)

# Predict
predict(object, progress_ne)

# Evaluate
object <- crossval(object, maxsam = 100, seed = 1)
plot_metrics(object)

# Summarize
summary(object)

# Report
report(object, name = "example_report", dir = getwd())
}
}
