% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{carousel}
\alias{carousel}
\title{Plot carousel}
\usage{
carousel(fun, x, ...)
}
\arguments{
\item{fun}{function. A plot function.}

\item{x}{S4 object. A PersephoneModel.}

\item{...}{extra arguments.}
}
\value{
Same as the function \code{slickR::slickR()}.
}
\description{
Plot carousel
}
