% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contrast.R
\name{contrast}
\alias{contrast}
\alias{contrast,PersephoneModelList-method}
\title{contrast models}
\usage{
contrast(object, ...)

\S4method{contrast}{PersephoneModelList}(
  object,
  by = "AICc",
  save = TRUE,
  name = "model_contrast",
  path = getwd(),
  ...
)
}
\arguments{
\item{object}{an object of class \code{PersephoneModelList}.}

\item{...}{extra arguments.}

\item{by}{character. Currently, only "AICc" is accepted.}

\item{save}{logical. Should the result be saved?}

\item{name}{character. The file name.}

\item{path}{character. The path to store the file.}
}
\value{
If \code{by = "aicc"}, a list of objects of class \code{aictab}.
}
\description{
Contrast models inside a \code{PersephoneModelList} object based on a
criterion of choice, such as AICc.
}
\examples{
\dontrun{
# Create a Region object
library(cronus)
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))

# Create a model
object1 <- new("PersephoneBin",
               region = region,
               crop = "Corn",
               data = progress_ne$Corn,
               formula = "CumPercentage ~ Time + agdd") # PersephoneModel

# Create another model
object2 <- new("PersephoneBin",
               region = region,
               crop = "Corn",
               data = progress_ne$Corn,
               formula = "Stage ~ Time + agdd + adayl") # PersephoneModel

# Concatenate the models
object <- c(object1, object2) # PersephoneModelList

# Fit
object <- fit(object)

# Contrast
contrast(object)
}
}
\seealso{
\code{\link[AICcmodavg:aictab]{AICcmodavg::aictab()}}
}
