% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{crossval}
\alias{crossval}
\alias{crossval,PersephoneModel-method}
\alias{crossval,PersephoneModelList-method}
\title{Cross - Validation}
\usage{
crossval(object, ...)

\S4method{crossval}{PersephoneModel}(object, test = 0.25, maxsam = 500, seed = 1)

\S4method{crossval}{PersephoneModelList}(object, ...)
}
\arguments{
\item{object}{an object of class \code{PersephoneModel} or \code{PersephoneModelList}.}

\item{...}{extra arguments.}

\item{test}{numeric. The proportion of seasons to be used for testing}

\item{maxsam}{numeric. The maximum number of Monte - Carlo samples to use.}

\item{seed}{numeric. The seed to be used in Monte - Carlo sampling.}
}
\value{
A list containing the calculated metrics.
}
\description{
Calculate the model performance metrics using leave-group-out (or Monte
Carlo) cross-validation.
}
\examples{
\dontrun{
# Create a Region object
library(cronus)
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))

# Create a model
object1 <- new("PersephoneBin",
               region = region,
               crop = "Corn",
               data = progress_ne$Corn,
               formula = "CumPercentage ~ Time + agdd") # PersephoneModel

# Create another model
object2 <- new("PersephoneCumLink",
               region = region,
               crop = "Soybeans",
               data = progress_ne$Soybeans,
               formula = "Stage ~ Time + agdd + adayl") # PersephoneModel

# Concatenate the models
object <- c(object1, object2) # PersephoneModelList

# Fit
object <- fit(object)

# Plot
plot(object, cumulative = TRUE, seasons = 2002)

# Predict
predict(object, progress_ne)

# Evaluate
object <- crossval(object, maxsam = 100, seed = 1)
plot_metrics(object)

# Summarize
summary(object)

# Report
report(object, name = "example_report", dir = getwd())
}
}
