% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossval.R
\name{crossval}
\alias{crossval}
\alias{crossval,PersephoneModel-method}
\alias{crossval,PersephoneModelList-method}
\title{Cross - Validation}
\usage{
crossval(object, ...)

\S4method{crossval}{PersephoneModel}(object, test = 0.25, maxsam = 500, seed = 1)

\S4method{crossval}{PersephoneModelList}(object, ...)
}
\arguments{
\item{object}{S4 object. The model of interest.}

\item{...}{extra arguments.}

\item{test}{numeric. The proportion of seasons to be used for testing}

\item{maxsam}{numeric. The maximum number of Monte - Carlo samples to use.}

\item{seed}{numeric. The seed to be used in Monte - Carlo sampling.}
}
\value{
A list.
}
\description{
Cross - Validation
}
\examples{
\dontrun{
mrmspe <- crossval(object, test = 0.25, maxsam = 100, seed = 2938)
}
}
