% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit}
\alias{fit}
\alias{fit,PersephoneModelList-method}
\alias{fit,PersephoneBinomial-method}
\alias{fit,PersephoneCumLink-method}
\title{Persephone model fitting}
\usage{
fit(object, ...)

\S4method{fit}{PersephoneModelList}(object, ...)

\S4method{fit}{PersephoneBinomial}(object, ...)

\S4method{fit}{PersephoneCumLink}(object, ...)
}
\arguments{
\item{object}{S4 object. The model of interest.}

\item{...}{extra arguments passed to update().}
}
\value{
S4 object. The model of interest.
}
\description{
Persephone model fitting
}
\examples{
\dontrun{
region <- c(name = "nebraska", type = "us state")
formula <- "Percentage ~ Time"
object <- new("PersephoneBinomial", region = region, formula = formula)
object <- fit(object, data = data)
}
}
