% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genTd.R
\name{genTd}
\alias{genTd}
\title{Generation of trading day regressors}
\usage{
genTd(freq = 12, fYear = 1960, lYear = 2099, hd, weight = rep(1,
  length(hd)))
}
\arguments{
\item{freq}{frequency of series}

\item{fYear}{first year of calculation of trading day regressor}

\item{lYear}{last year of calculation of trading day regressor}

\item{hd}{holidays, list of 1. exact date ("mm-dd") and/or 2. name of day
(listHolidays() from timeDate function) and/or
3. easter relation (e.g. "easter+39", "easter-3")}

\item{weight}{vector of individual weights for each holiday with length of hd}
}
\value{
list with tree list elements. 1. matrix of trading day counts for each
individual day, month and year. Holidays which are delivered through the
parameter hd are assigned to the number of sundays. 2. multiple time
series object with 6 non-centered trading day regressors plus 1
non-centered working day regressor (columns 1 to 6 refer to Mondays to
Saturdays, colunm 7 refers to the one-regressor case of 5 working days).
In case of trading day adjustment the first 6 columns have to be selected,
in case of wording day adjustment only column 7 has to be selected.
3. multiple time series object with 6 centered trading day regressors
plus 1 centered working day regressor
}
\description{
User defined trading day regressors are generated depending on a list of
country-specific holidays and associated weights. Each holiday can be given
a specific weight depending on the extent of the work load of that specific holiday.
}
\examples{
\dontrun{
hdAT <- genTd(hd = list("NewYearsDay","Epiphany","EasterMonday","LaborDay",
                        "PentecostMonday","Ascension","CorpusChristi",
                        "AssumptionOfMary","10-26","AllSaints",
                        "ITImmaculateConception","ChristmasEve","ChristmasDay",
                        "BoxingDay","12-31"))
# Representation of object with alternative weights for trading days
hdAT1 <- genTd(hd = list("NewYearsDay","Epiphany","EasterMonday","LaborDay",
                         "PentecostMonday","Ascension","CorpusChristi",
                         "AssumptionOfMary","10-26","AllSaints",
                         "ITImmaculateConception","ChristmasEve","ChristmasDay",
                         "BoxingDay","12-31"),
                        weight = c(rep(1,11),0.6,rep(1,2),0.6))
obj_x13 <- per_x13(AirPassengers, template = "RSA3", tradingdays.option = "UserDefined",
                        usrdef.varType = "Calendar",
                        usrdef.varEnabled = TRUE, usrdef.var = hdAT[[4]][,1:6])
obj_x13$run()
obj_x13$output$regarima
}
}
