% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_class}
\alias{get_class}
\alias{get_class,ProgressModel-method}
\alias{get_class,ProgressModelList-method}
\alias{get_region}
\alias{get_region,ProgressModel-method}
\alias{get_region,ProgressModelList-method}
\alias{get_label}
\alias{get_label,ProgressModel-method}
\alias{get_label,ProgressModelList-method}
\alias{get_formula}
\alias{get_formula,ProgressModel-method}
\alias{get_formula,ProgressModelList-method}
\alias{get_model}
\alias{get_model,ProgressModel-method}
\alias{get_model,ProgressModelList-method}
\alias{get_crops}
\alias{get_crops,ProgressModel-method}
\alias{get_crops,ProgressModelList-method}
\alias{get_index}
\alias{get_index,ProgressModelList-method}
\alias{get_stages}
\alias{get_stages,ProgressModel-method}
\alias{get_stages,ProgressModelList-method}
\alias{get_seasons}
\alias{get_seasons,ProgressModel-method}
\alias{get_seasons,ProgressModelList-method}
\alias{get_link}
\alias{get_link,ProgressModel-method}
\alias{get_link,ProgressModelList-method}
\title{Get object slots}
\usage{
get_class(object, ...)

\S4method{get_class}{ProgressModel}(object, ...)

\S4method{get_class}{ProgressModelList}(object, ...)

get_region(object, ...)

\S4method{get_region}{ProgressModel}(object, ...)

\S4method{get_region}{ProgressModelList}(object, ...)

get_label(object, ...)

\S4method{get_label}{ProgressModel}(object, ...)

\S4method{get_label}{ProgressModelList}(object, ...)

get_formula(object, ...)

\S4method{get_formula}{ProgressModel}(object, ...)

\S4method{get_formula}{ProgressModelList}(object, ...)

get_model(object, ...)

\S4method{get_model}{ProgressModel}(object, ...)

\S4method{get_model}{ProgressModelList}(object, ...)

get_crops(object, ...)

\S4method{get_crops}{ProgressModel}(object)

\S4method{get_crops}{ProgressModelList}(object)

get_index(object, ...)

\S4method{get_index}{ProgressModelList}(object, crop = NULL)

get_stages(object, ...)

\S4method{get_stages}{ProgressModel}(object)

\S4method{get_stages}{ProgressModelList}(object)

get_seasons(object, ...)

\S4method{get_seasons}{ProgressModel}(object)

\S4method{get_seasons}{ProgressModelList}(object)

get_link(object, ...)

\S4method{get_link}{ProgressModel}(object)

\S4method{get_link}{ProgressModelList}(object)
}
\arguments{
\item{object}{an object of class \code{ProgressModel} or \code{ProgressModelList}.}

\item{...}{extra arguments}

\item{crop}{character. A crop of choice.}
}
\value{
character. The \code{class()} of the \code{object}.
}
\description{
Get object slots
}
\section{Functions}{
\itemize{
\item \code{get_region()}: returns a character. The \code{region} slot of the \code{object}.

\item \code{get_label()}: returns a character. The \code{label} slot of the \code{object}.

\item \code{get_formula()}: returns a character. The \code{formula} slot of the \code{object}.

\item \code{get_model()}: returns a character. The \code{model} slot of the \code{object}.

\item \code{get_crops()}: returns a character. The \code{crop} slot of the \code{object}.

\item \code{get_index()}: returns a numeric The indices that correspond to the
\code{crop} slot of the \code{object}.

\item \code{get_stages()}: returns a factor. The \code{stages} included in \code{data}.

\item \code{get_seasons()}: returns a numeric. The \code{seasons} included in \code{data}.

\item \code{get_link()}: returns a character The \code{link} slot of the \code{object}.

}}
\examples{
\dontrun{
# Create a Region object
library(cronus)
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))

# Create a model
object <- new("ProgressBM",
              label = "Base_logit",
              region = region,
              crop = "Corn",
              data = progress_ne$Corn,
              link = "logit",
              formula = "CumPercentage ~ Time + agdd + adayl") # ProgressModel

# Fit
object <- fit(object)

# Get slots
get_class(object)
get_label(object)
get_formula(object)
get_model(object)
get_crops(object)
get_region(object)
get_index(object, crop = "Corn")
get_stages(object)
get_seasons(object)
}
}
