% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalTimeSeries.R
\docType{data}
\name{hierarchicalTimeSeries}
\alias{hierarchicalTimeSeries}
\title{Define a hierarchical time series}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
hierarchicalTimeSeries
}
\description{
Combine mutliple objects of persephone objects into a new persephone object.
The resulting time series can perform direct and indirect adjustments.
}
\section{Inherits}{
 \link{persephone}
}

\section{Constructor}{

\preformatted{
hierarchicalTimeSeries$new(...)
}
\itemize{
\item \code{...} One or more \code{persephone} objects that use the same time instances.
Must be named
}
}

\examples{
obj_x13 <- x13Single$new(AirPassengers, "RSA3")

ht <- hierarchicalTimeSeries$new(obj_x13, obj_x13)
ht$run()
ht$adjusted
ht$adjusted_indirect

ht2 <- hierarchicalTimeSeries$new(ht, obj_x13)
ht2$run()
ht2$adjusted
ht2$adjusted_indirect

}
\keyword{datasets}
