% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchicalTimeSeries.R
\docType{data}
\name{hierarchicalTimeSeries}
\alias{hierarchicalTimeSeries}
\alias{per_hts}
\title{Define a hierarchical time series}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
hierarchicalTimeSeries
}
\description{
Combine mutliple objects of persephone objects into a new persephone object.
The resulting time series can perform direct and indirect adjustments.
}
\section{Inherits}{
 \link{persephone}
}

\section{Constructor}{

\preformatted{
per_hts(...)
}
\itemize{
\item \code{...} should contain one or more \code{persephone} objects that use the same
time instances. All elements supplied here must be named.
\item \code{model} specifies a model to be used. tramoseats or x13
\item \code{userdefined} is passed as the userdefined argument to \code{tramoseats} or
\code{x13}
\item \code{spec} a model specification returned by \code{\link[=x13_spec]{x13_spec()}} or
\code{\link[=tramoseats_spec]{tramoseats_spec()}}
}
}

\examples{
obj_x13 <- per_x13(AirPassengers, "RSA3")

ht <- per_hts(a = obj_x13, b = obj_x13, model = "x13")
ht$run()
ht$adjusted
ht$adjusted_indirect

ht2 <- per_hts(a = ht, b = obj_x13)
ht2$run()
ht2$adjusted
ht2$adjusted_indirect

data(ipi_c_eu,package = "RJDemetra")
# Reducing the data set to the EU28 countries
ipi_eu <- ipi_c_eu[,-c(1:3,32:37)]
# We now build persephone objects for the countries
# (the lowest level in the hierachy)
ts_28 <- lapply(ipi_eu, per_tramo)
# We want to add an extra layer and split the EU28 countries in two groups
ht_half_europe_1 <- do.call(per_hts,ts_28[1:14],)
# possible future syntax
#per_hts(ts_28[1:14], model=list(AT="x13",BE="tramoseats"),spec=list(),
#outlier_enabled(..=TRUE),tradings.leapyw=list(LV=TRUE,AT=dfdjkl,))
ht_half_europe_2 <- do.call(per_hts,ts_28[15:28])
# Now we generate the object for EU28
 ht_europe <- per_hts(
  halfEU_1 = ht_half_europe_1,
  halfEU_2 = ht_half_europe_2,
  spec="RSA5c",model="x13")
# start the seasonal adjustment
ht_europe$run()
# accessing the directly and indirectly adjusted series for EU28
ht_europe$adjusted
ht_europe$adjusted_indirect
# accessing the directly and indirectly adjusted series for half of Europe
ht_europe$components$halfEU_2$adjusted
ht_europe$components$halfEU_2$adjusted_indirect
# accessing the adjusted series for a country
ht_europe$components$halfEU_2$components$AT$adjusted
}
\keyword{datasets}
