% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.persephone.R
\name{is.persephone}
\alias{is.persephone}
\title{is.persephone}
\usage{
is.persephone(x)
}
\description{
Check if an R object is of or derives from the class \code{persephone}
}
\examples{
is.persephone(2)
}
