% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate.R
\name{iterate}
\alias{iterate}
\title{Iterate over a hierarchical persephone object}
\usage{
iterate(x, fun, asTable = FALSE, component = "", unnest = FALSE)
}
\arguments{
\item{x}{a hierarchical persephone object}

\item{fun}{a function with one argument}

\item{asTable}{should the return value be coerced to a \code{data.frame}?}

\item{component}{a component id from which the iteration should start}

\item{unnest}{should the return value be unnested, so that every component
is an entry of the resulting list?}
}
\description{
This funtion is an alias to the method \code{iterate} and can be used to iterate
over all children of a hierarchical time series produced with \code{perHts()}.
}
\examples{
obj_x13 <- perX13(AirPassengers, "rsa3")
ht <- perHts(a = obj_x13, b = obj_x13, method = "x13")
ht2 <- perHts(a = ht, b = obj_x13)
ht2$iterate(function(x) {class(x)[1]})
ht2$iterate(function(x) {list(class = class(x)[1])}, asTable = TRUE)
ht2$iterate(function(x) {list(class = class(x)[1])}, unnest = TRUE)
}
