% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_hts.R
\name{hierarchicalTimeSeries}
\alias{hierarchicalTimeSeries}
\title{Define a hierarchical time series}
\description{
Combine mutliple objects of persephone objects into a new persephone object.
The resulting time series can perform direct and indirect adjustments.
}
\section{Inherits}{
 \link{persephone}
}

\section{Constructor}{

\preformatted{
per_hts(..., method = c("tramoseats", "x13"),
        userdefined = NULL, spec = NULL)
}
\itemize{
\item \code{...} should contain one or more objects which are either of class
\code{persephone} or can be coerced to \code{persephone} objects with
\code{as_persephone}. If more than one element is supplied, the underlying
time series must have the same time instances. All elements supplied
in \code{...} must be named.
\item \code{list} a list of \code{persephone} objects as alternative input to \code{...}. This
argument can also handle \verb{mts objects}
\item \code{weights} either a vector
if the same weight is used for all time points or a list of ts objects or a
mts object if the weight varies for different time points. They must have
the same length as the number of components.
\item \code{method} specifies the method to be used for the direct adjustment of the
aggregate series. tramoseats or x13
\item \code{userdefined} is passed as the userdefined argument to \code{\link[=tramoseats]{tramoseats()}} or
\code{\link[=x13]{x13()}}
\item \code{spec} a model specification returned by \code{\link[=x13_spec]{x13_spec()}} or
\code{\link[=tramoseats_spec]{tramoseats_spec()}}
}
}

\section{Fields}{

\itemize{
\item \verb{$components}. A list of \code{persephone} objects.
\item \strong{\verb{$adjusted_indirect}}. Results from indirect adjustments which means the
components are first adjusted and the adjusted series are then aggregated
}
}

\section{Methods}{

\itemize{
\item \verb{$iterate(fun, ...)} can be used to iterate over the hierarchy tree. See
\code{\link[=iterate]{iterate()}}.
\item \verb{$set_options(userdefined, spec, recursive = TRUE, component = "")} sets
options for all entries of the dependency tree recursively if
\code{recursive = TRUE} (the default). See
\code{vignette("persephone-hierarchical")}.
}
}

\examples{
obj_x13 <- per_x13(AirPassengers, "RSA3")

ht <- per_hts(a = obj_x13, b = obj_x13, method = "x13")
ht$run()
ht$adjusted_direct
ht$adjusted_indirect

ht2 <- per_hts(a = ht, b = obj_x13)
ht2$run()
ht2$adjusted_direct
ht2$adjusted_indirect

#-------- example with industrial price indices -----------

data(ipi_c_eu, package = "RJDemetra")
# Reducing the data set to the EU28 countries
ipi_eu <- ipi_c_eu[, -c(1:3, 32:37)]

# We want to add an extra layer and split the EU28 countries in two groups
ht_half_europe_1  <- per_hts(list = ipi_eu[, 1:14], method = "tramoseats")

# Alternative way to use do.call
ht_half_europe_2 <- do.call(per_hts, as.list(ipi_eu[, 15:28]))

# Now we generate the object for EU28
ht_europe <- per_hts(
  halfEU_1 = ht_half_europe_1,
  halfEU_2 = ht_half_europe_2,
  spec = "RSA5c",
  method = "x13"
)

# start the seasonal adjustment
ht_europe$run()

# accessing the directly and indirectly adjusted series for EU28
ht_europe$adjusted_direct
ht_europe$adjusted_indirect

# accessing the directly and indirectly adjusted series for half of Europe
ht_europe$components$halfEU_2$adjusted_direct
ht_europe$components$halfEU_2$adjusted_indirect

# accessing the adjusted series for a country
ht_europe$components$halfEU_2$components$AT$adjusted_direct
}
\section{Super class}{
\code{\link[persephone:persephone]{persephone::persephone}} -> \code{hierarchicalTimeSeries}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{hierarchicalTimeSeries$new()}}
\item \href{#method-run}{\code{hierarchicalTimeSeries$run()}}
\item \href{#method-print}{\code{hierarchicalTimeSeries$print()}}
\item \href{#method-set_options}{\code{hierarchicalTimeSeries$set_options()}}
\item \href{#method-iterate}{\code{hierarchicalTimeSeries$iterate()}}
\item \href{#method-get_component}{\code{hierarchicalTimeSeries$get_component()}}
\item \href{#method-generate_qr_table}{\code{hierarchicalTimeSeries$generate_qr_table()}}
\item \href{#method-clone}{\code{hierarchicalTimeSeries$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="plot">}\href{../../persephone/html/persephone.html#method-plot}{\code{persephone::persephone$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="plotResiduals">}\href{../../persephone/html/persephone.html#method-plotResiduals}{\code{persephone::persephone$plotResiduals()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="plotSeasIrrCal">}\href{../../persephone/html/persephone.html#method-plotSeasIrrCal}{\code{persephone::persephone$plotSeasIrrCal()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="updateParams">}\href{../../persephone/html/persephone.html#method-updateParams}{\code{persephone::persephone$updateParams()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$new(
  ...,
  method = c("tramoseats", "x13"),
  userdefined = NULL,
  spec = NULL,
  list = NULL,
  weights = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\subsection{Method \code{run()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$run(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_options"></a>}}
\subsection{Method \code{set_options()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$set_options(
  userdefined = NA,
  spec = NA,
  recursive = TRUE,
  component = ""
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-iterate"></a>}}
\subsection{Method \code{iterate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$iterate(
  fun,
  as_table = FALSE,
  component = "",
  unnest = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_component"></a>}}
\subsection{Method \code{get_component()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$get_component(component_id)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-generate_qr_table"></a>}}
\subsection{Method \code{generate_qr_table()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$generate_qr_table(component = "")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{hierarchicalTimeSeries$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
