% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persephone-package.R
\docType{package}
\name{persephone-package}
\alias{persephone}
\alias{persephone-package}
\title{persephone: Remote Sensing Applications in Crop Phenology}
\description{
This package is part of the prometheus family of R packages.
}
\author{
\strong{Maintainer}: Ioannis Oikonomidis \email{goikon@math.uoa.gr} (\href{https://orcid.org/0000-0001-8130-2104}{ORCID})

Authors:
\itemize{
  \item Samis Trevezas \email{strevezas@math.uoa.gr} [thesis advisor]
}

}
\keyword{internal}
