% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persephone.R
\name{persephone}
\alias{persephone}
\title{persephone class}
\description{
Common interface for x13 and tramoseats
}
\section{Constructor}{

\itemize{
\item \code{new()}: Initializes an object. See the documentation for the derived
classes for details.
}
}

\section{Methods}{

\itemize{
\item \code{run(verbose = FALSE)}: Runs the model
\item \code{updateParams(...)}: Updates the current parameters
\item \code{plot(...)}: Interactive time series plot for a persephone object, see
\link{plot.persephoneSingle} or \link{plot.hierarchicalTimeSeries}
\item \code{plotSeasIrrCal(...)}: Interactive plot of the seasonal component,
irregular component and calendar
effects for a persephone object, see \link{plotSeasIrrCal}
\item \code{plotResiduals(...)}: Several interactive plots in connection with
residuals for a persephone
object, see \link{plotResiduals}
}
}

\section{Read-Only fields}{

\itemize{
\item \code{params}: A parameters object of class \code{SA_spec}. See \link{x13_spec} and
\link{tramoseats_spec}.
\item \code{ts}: The time series.
\item \code{output}: The return value from the underlying functions \code{x13_fast} or
\code{tramoseats_fast}. This slot will
be empty (\code{NULL}) before \code{run()} is called for the first time.
}
}

\section{Parameters}{

\itemize{
\item \code{verbose}: should the output from the underlying rjd3x13/rjd3tramoseats
function be displayed after the
run? If \code{FALSE} (the default), the output will be returned invisibly.
}
}

\section{derived classes}{

\code{\link[=perX13]{perX13()}}, \code{\link[=perTramo]{perTramo()}}, \code{\link[=perHts]{perHts()}}
}

\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{params}}{A parameters object of class \code{SA_spec}. See \code{\link[=x13_spec]{x13_spec()}}
and \code{\link[=tramoseats_spec]{tramoseats_spec()}}.}

\item{\code{ts}}{the (unajusted) time series}

\item{\code{tsp}}{the \code{\link[=tsp]{tsp()}} of the underlying time series}

\item{\code{output}}{The return value from the underlying functions x13_fast() or
tramoseats_fast(). This slot will be empty (NULL) before run() is called for
the first time.}

\item{\code{adjusted}}{get the adjusted series}

\item{\code{adjustedDirect}}{see \link{perHts}}

\item{\code{spec}}{specifications passed to \code{\link[=x13_fast]{x13_fast()}} and \code{\link[=tramoseats_fast]{tramoseats_fast()}} when the
\verb{$run()} method is invoked}

\item{\code{forecasts}}{get forecasts from the model}

\item{\code{forecastsDirect}}{get direct forecasts from the model}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-persephone-new}{\code{persephone$new()}}
\item \href{#method-persephone-run}{\code{persephone$run()}}
\item \href{#method-persephone-updateParams}{\code{persephone$updateParams()}}
\item \href{#method-persephone-plot}{\code{persephone$plot()}}
\item \href{#method-persephone-plotResiduals}{\code{persephone$plotResiduals()}}
\item \href{#method-persephone-plotSeasIrrCal}{\code{persephone$plotSeasIrrCal()}}
\item \href{#method-persephone-plotSiRatios}{\code{persephone$plotSiRatios()}}
\item \href{#method-persephone-plotSpectrum}{\code{persephone$plotSpectrum()}}
\item \href{#method-persephone-print}{\code{persephone$print()}}
\item \href{#method-persephone-iterate}{\code{persephone$iterate()}}
\item \href{#method-persephone-generateQrTable}{\code{persephone$generateQrTable()}}
\item \href{#method-persephone-setOptions}{\code{persephone$setOptions()}}
\item \href{#method-persephone-fixModel}{\code{persephone$fixModel()}}
\item \href{#method-persephone-fixOutlier}{\code{persephone$fixOutlier()}}
\item \href{#method-persephone-clone}{\code{persephone$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-new"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-new}{}}}
\subsection{Method \code{new()}}{
create a new persephone object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$new()}\if{html}{\out{</div>}}
}

\subsection{Details}{
this function should not be used directly because this class is
abstract. Only initialize derived classes.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-run"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-run}{}}}
\subsection{Method \code{run()}}{
perform a run of the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$run(verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{whether to show outputs}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-updateParams"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-updateParams}{}}}
\subsection{Method \code{updateParams()}}{
update parameters for the adjustment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$updateParams(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to \code{x13_spec()} or \code{tramoseats_spec()}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-plot"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-plot}{}}}
\subsection{Method \code{plot()}}{
visualize the results of an adjustment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$plot(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to \code{\link[=plot]{plot()}}
see \code{\link[=plot.persephoneSingle]{plot.persephoneSingle()}} or \code{\link[=plot.hierarchicalTimeSeries]{plot.hierarchicalTimeSeries()}}
#' @description visualize the results of an adjustment
#' @param ...  passed to \code{\link[=plot]{plot()}}
#'   see \code{\link[=plot.persephoneSingle]{plot.persephoneSingle()}} or \code{\link[=plot.hierarchicalTimeSeries]{plot.hierarchicalTimeSeries()}}
plot = function(...) {
plot(self, ...)
},}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-plotResiduals"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-plotResiduals}{}}}
\subsection{Method \code{plotResiduals()}}{
visualize residuals
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$plotResiduals(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to \code{\link[=plotResiduals]{plotResiduals()}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-plotSeasIrrCal"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-plotSeasIrrCal}{}}}
\subsection{Method \code{plotSeasIrrCal()}}{
visualize the seasonal component, irregular
component and calendar effects for a persephone object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$plotSeasIrrCal(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to \code{\link[=plotSeasIrrCal]{plotSeasIrrCal()}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-plotSiRatios"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-plotSiRatios}{}}}
\subsection{Method \code{plotSiRatios()}}{
visualize SI-ratios (and, in case of x11, of seasonal factors) by
period
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$plotSiRatios(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to \code{\link[=plotSiRatios]{plotSiRatios()}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-plotSpectrum"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-plotSpectrum}{}}}
\subsection{Method \code{plotSpectrum()}}{
visualize the periodogram or the autogregressive spectrum.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$plotSpectrum(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{passed to \code{\link[=plotSpectrum]{plotSpectrum()}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-print"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-print}{}}}
\subsection{Method \code{print()}}{
display a persephone object in the console
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-iterate"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-iterate}{}}}
\subsection{Method \code{iterate()}}{
see \link{perHts}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$iterate(fun, asTable = FALSE, unnest = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{see \link{perHts}}

\item{\code{asTable}}{see \link{perHts}}

\item{\code{unnest}}{see \link{perHts}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-generateQrTable"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-generateQrTable}{}}}
\subsection{Method \code{generateQrTable()}}{
create a table for the eurostat quality report
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$generateQrTable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-setOptions"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-setOptions}{}}}
\subsection{Method \code{setOptions()}}{
update options for the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$setOptions(userdefined = NA, spec = NA, recursive = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{userdefined}}{see \code{\link[=x13_fast]{x13_fast()}} and \code{\link[=tramoseats_fast]{tramoseats_fast()}}}

\item{\code{spec}}{see \code{\link[=x13_fast]{x13_fast()}} and \code{\link[=tramoseats_fast]{tramoseats_fast()}}}

\item{\code{recursive}}{only applicable to hierarchical series. propagates
the updates to sub-series. see \link{perHts}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-fixModel"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-fixModel}{}}}
\subsection{Method \code{fixModel()}}{
fix the arima model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$fixModel(verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{if TRUE the changed parameters will be reported}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-fixOutlier"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-fixOutlier}{}}}
\subsection{Method \code{fixOutlier()}}{
create a new single object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$fixOutlier(timespan = 12, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{timespan}}{number of months from the end of the time series
where outliers are not fixed}

\item{\code{verbose}}{if TRUE the changed parameters will be reported}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-persephone-clone"></a>}}
\if{latex}{\out{\hypertarget{method-persephone-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephone$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
