% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persephone.R
\docType{data}
\name{persephone}
\alias{persephone}
\title{persephone class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
persephone
}
\description{
Common interface for x13 and tramoseats
}
\section{Methods}{

\itemize{
\item \code{run()}: Runs the model
\item \code{updateParams(...)}: Updates the current parameters
}
}

\keyword{datasets}
