% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persephone.R
\name{persephone}
\alias{persephone}
\title{persephone class}
\description{
Common interface for x13 and tramoseats
}
\section{Constructor}{

\itemize{
\item \code{new()}: Initializes an object. See the documentation for the derived
classes for details.
}
}

\section{Methods}{

\itemize{
\item \code{run(verbose = FALSE)}: Runs the model
\item \code{updateParams(...)}: Updates the current parameters
\item \code{plot(...)}: Interactive time series plot for a persephone object, see
\link{plot.persephone}
\item \code{plotSeasIrrCal(...)}: Interactive plot of the seasonal component,
irregular component and calendar
effects for a persephone object, see \link{plotSeasIrrCal}
\item \code{plotResiduals(...)}: Several interactive plots in connection with
residuals for a persephone
object, see \link{plotResiduals}
}
}

\section{Read-Only fields}{

\itemize{
\item \code{params}: A parameters object of class \code{SA_spec}. See \link{x13_spec} and
\link{tramoseats_spec}.
\item \code{ts}: The time series.
\item \code{output}: The return value from the underlying functions \code{x13} or
\code{tramoseats}. This slot will
be empty (\code{NULL}) before \code{run()} is called for the first time.
}
}

\section{Parameters}{

\itemize{
\item \code{verbose}: should the output from the underlying RJDemetra function be
displayed after the
run? If \code{FALSE} (the default), the output will be returned invisibly.
}
}

\seealso{
\link{x13Single}, \link{tramoseatsSingle}
}
