% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_single.R
\name{persephoneSingle}
\alias{persephoneSingle}
\title{Class for single time series}
\description{
Class for single time series

Class for single time series
}
\details{
A common class for the wrapper classes \link{x13Single} and \link{tramoseatsSingle}
}
\section{Super class}{
\code{\link[persephone:persephone]{persephone::persephone}} -> \code{persephoneSingle}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{persephoneSingle$new()}}
\item \href{#method-run}{\code{persephoneSingle$run()}}
\item \href{#method-clone}{\code{persephoneSingle$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="generate_qr_table">}\href{../../persephone/html/persephone.html#method-generate_qr_table}{\code{persephone::persephone$generate_qr_table()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="iterate">}\href{../../persephone/html/persephone.html#method-iterate}{\code{persephone::persephone$iterate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="plot">}\href{../../persephone/html/persephone.html#method-plot}{\code{persephone::persephone$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="plotResiduals">}\href{../../persephone/html/persephone.html#method-plotResiduals}{\code{persephone::persephone$plotResiduals()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="plotSeasIrrCal">}\href{../../persephone/html/persephone.html#method-plotSeasIrrCal}{\code{persephone::persephone$plotSeasIrrCal()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="print">}\href{../../persephone/html/persephone.html#method-print}{\code{persephone::persephone$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="set_options">}\href{../../persephone/html/persephone.html#method-set_options}{\code{persephone::persephone$set_options()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="updateParams">}\href{../../persephone/html/persephone.html#method-updateParams}{\code{persephone::persephone$updateParams()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
create a new single object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephoneSingle$new(ts, template = "RSA0", userdefined = NULL, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ts}}{a time series object}

\item{\code{template}}{passed as the \code{spec} argument of \code{\link[=x13]{x13()}} or
\code{\link[=tramoseats]{tramoseats()}}}

\item{\code{userdefined}}{passed as the \code{userdefined} argument of \code{\link[=x13]{x13()}} or
\code{\link[=tramoseats]{tramoseats()}}}

\item{\code{...}}{passed to \code{\link[=x13_spec]{x13_spec()}} or \code{\link[=tramoseats_spec]{tramoseats_spec()}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run"></a>}}
\subsection{Method \code{run()}}{
run the model
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephoneSingle$run(verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{wether to display outputs after the run}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{persephoneSingle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
