% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot}
\alias{plot}
\alias{plot,PersephoneModel,ANY-method}
\alias{plot,PersephoneModelList,ANY-method}
\title{Plot}
\usage{
plot(x, y, ...)

\S4method{plot}{PersephoneModel,ANY}(
  x,
  seasons = 2002,
  cumulative = TRUE,
  save = FALSE,
  dir = getwd(),
  file = "plot.pdf",
  width = 15,
  height = 8
)

\S4method{plot}{PersephoneModelList,ANY}(
  x,
  ncol = NULL,
  nrow = NULL,
  save = FALSE,
  dir = getwd(),
  file = "plot.pdf",
  width = 15,
  height = 8,
  ...
)
}
\arguments{
\item{x}{an S4 object. The model of interest.}

\item{y}{not used.}

\item{...}{extra arguments}

\item{seasons}{numeric. The seasons of choice to plot.}

\item{cumulative}{logical. Should the cumulative percentages be plotted?}

\item{save}{logical. Should the plot be saved?}

\item{dir}{character. The directory in which the plot will be saved.}

\item{file}{character. The file name.}

\item{width}{numeric. The width of the plot in inches.}

\item{height}{numeric. The height of the plot in inches.}

\item{ncol}{numeric. Number of rows in the plot layout.}

\item{nrow}{numeric. Number of columns in the plot layout.}
}
\value{
A plot, created with \code{ggplot2} and \code{ggpubr}. The plot is
returned wrapped inside \code{invisible()}.
}
\description{
Plot
}
\examples{
\dontrun{
# Create a Region object
library(cronus)
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))

# Create a model
object1 <- new("PersephoneBin",
               region = region,
               crop = "Corn",
               data = progress_ne$Corn,
               formula = "CumPercentage ~ Time + agdd") # PersephoneModel

# Create another model
object2 <- new("PersephoneCumLink",
               region = region,
               crop = "Soybeans",
               data = progress_ne$Soybeans,
               formula = "Stage ~ Time + agdd + adayl") # PersephoneModel

# Concatenate the models
object <- c(object1, object2) # PersephoneModelList

# Fit
object <- fit(object)

# Plot
plot(object, cumulative = TRUE, seasons = 2002)

# Predict
predict(object, progress_ne)

# Evaluate
object <- crossval(object, maxsam = 100, seed = 1)
plot_metrics(object)

# Summarize
summary(object)

# Report
report(object, name = "example_report", dir = getwd())
}
}
