% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_metrics}
\alias{plot_metrics}
\alias{plot_metrics,PersephoneModel-method}
\alias{plot_metrics,PersephoneModelList-method}
\title{Plot Metrics}
\usage{
plot_metrics(object, ...)

\S4method{plot_metrics}{PersephoneModel}(
  object,
  title = NULL,
  save = FALSE,
  dir = getwd(),
  file = "plot.pdf",
  width = 15,
  height = 8,
  print_plots = TRUE
)

\S4method{plot_metrics}{PersephoneModelList}(
  object,
  ncol = NULL,
  nrow = NULL,
  save = FALSE,
  dir = getwd(),
  file = "plot.pdf",
  width = 15,
  height = 8,
  ...
)
}
\arguments{
\item{object}{an S4 object. The model of interest.}

\item{...}{extra arguments.}

\item{title}{character. The plot title.}

\item{save}{logical. Should the plot be saved?}

\item{dir}{character. The directory in which the plot will be saved.}

\item{file}{character. The file name.}

\item{width}{numeric. The width of the plot in inches.}

\item{height}{numeric. The height of the plot in inches.}

\item{print_plots}{logical. Should the plots be printed?}

\item{ncol}{numeric. Number of rows in the plot layout.}

\item{nrow}{numeric. Number of columns in the plot layout.}
}
\value{
A plot, created with \code{ggplot2} and \code{ggpubr}. The plot is
returned wrapped inside \code{invisible()}.
}
\description{
Create plots of the model metrics.
}
\examples{
\dontrun{
# Create a Region object
library(cronus)
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))

# Create a model
object1 <- new("PersephoneQuasiBin",
             region = region,
            crop = "Corn",
            data = progress_ne$Corn,
            formula = "Stage ~ Time + agdd") # PersephoneModel

# Create another model
object2 <- new("PersephoneCumLink",
            region = region,
            crop = "Soybeans",
            data = progress_ne$Soybeans,
            formula = "Stage ~ Time + agdd + adayl") # PersephoneModel

# Concatenate the models
object <- c(object1, object2) # PersephoneModelList

# Fit
object <- fit(object)

# Plot
plot(object, cumulative = TRUE, seasons = 2002)

# Predict
predict(object, progress_ne)

# Evaluate
object <- crossval(object, maxsam = 100, seed = 1)
plot_metrics(object)

# Summarize
summary(object)

# Report
report(object, name = "example_report", dir = getwd())
}
}
