% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResiduals.R
\name{plotResiduals}
\alias{plotResiduals}
\title{Several interactive plots in connection with residuals for a persephone object}
\usage{
plotResiduals(object, which = c("res", "acf", "acf2", "pacf", "sreshist",
  "nqq"), main = NULL, plotly = TRUE, ...)
}
\arguments{
\item{object}{an object of class \code{\link{persephone}}.}

\item{which}{character selecting the preferred type of plot (\code{"res"},\code{"acf"},\code{"acf2"},
\code{"pacf"},\code{"sreshist"},\code{"nqq"}), see Details.}

\item{main}{..}

\item{plotly}{..}

\item{...}{..}
}
\value{
Returns an object of class \code{dygraphs} or of class \code{ggplot}.
}
\description{
Produces either a dygraph (see the \href{https://rstudio.github.io/dygraphs/}{online documentation}
for more detail) or a ggplot object for objects of class \code{persephone}.
}
\details{
\describe{
\item{res:}{residuals}
\item{acf:}{autocorrelations of the residuals}
\item{pacf:}{partial autocorrelations of the residuals}
\item{acf2:}{autocorrelations of the squared residuals}
\item{sreshist:}{histogram of standardized residuals including normal curve}
\item{nqq:}{normal q-q plot of standardized residuals}
}
}
\examples{
data(AirPassengers, package = "datasets")
# Generate a persephone object, in this case an x13Single object
obj <- x13Single$new(AirPassengers, "RSA1")
## userdefined noch zu x13Single dazugeben

obj$run()
# Plot the residuals after run
plotResiduals(obj)
# Plot the autocorrelations of the residuals
plotResiduals(obj, which="acf")

## TO DO:
# maybe other theme_bw()
# maybe one standard for colors (for dygraph and ggplots -> take dygraph colors for ggplots?)
# maybe adjust tooltip for ggplotly (e.g. when count is shown even though y=density)

}
