% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\alias{predict,PersephoneModelList-method}
\alias{predict,PersephoneBin-method}
\alias{predict,PersephoneMixedBin-method}
\alias{predict,PersephoneCumLink-method}
\alias{predict,PersephoneMixedCumLink-method}
\title{\code{persephone} model prediction}
\usage{
predict(object, ...)

\S4method{predict}{PersephoneModelList}(object, pdata)

\S4method{predict}{PersephoneBin}(object, pdata)

\S4method{predict}{PersephoneMixedBin}(object, pdata)

\S4method{predict}{PersephoneCumLink}(object, pdata)

\S4method{predict}{PersephoneMixedCumLink}(object, pdata)
}
\arguments{
\item{object}{an object of class \code{PersephoneModel} or \code{PersephoneModelList}.}

\item{...}{extra arguments.}

\item{pdata}{data.frame. The data used in model prediction.}
}
\value{
data.frame.
}
\description{
Calculate the predicted values of a model. The \code{object}, provided needs to
have already passed from \code{fit}, in order to include a \code{model} slot.
}
\examples{
\dontrun{
# Create a Region object
library(cronus)
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))

# Create a model
object1 <- new("PersephoneBin",
               region = region,
               crop = "Corn",
               data = progress_ne$Corn,
               formula = "CumPercentage ~ Time + agdd") # PersephoneModel

# Create another model
object2 <- new("PersephoneCumLink",
               region = region,
               crop = "Soybeans",
               data = progress_ne$Soybeans,
               formula = "Stage ~ Time + agdd + adayl") # PersephoneModel

# Concatenate the models
object <- c(object1, object2) # PersephoneModelList

# Fit
object <- fit(object)

# Plot
plot(object, cumulative = TRUE, seasons = 2002)

# Predict
predict(object, progress_ne)

# Evaluate
object <- crossval(object, maxsam = 100, seed = 1)
plot_metrics(object)

# Summarize
summary(object)

# Report
report(object, name = "example_report", dir = getwd())
}
}
