% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\alias{predict,PersephoneModelList-method}
\alias{predict,PersephoneBinomial-method}
\alias{predict,PersephoneCumLink-method}
\title{Persephone model prediction}
\usage{
predict(object, ...)

\S4method{predict}{PersephoneModelList}(object, pdata)

\S4method{predict}{PersephoneBinomial}(object, pdata)

\S4method{predict}{PersephoneCumLink}(object, pdata)
}
\arguments{
\item{object}{S4 object. The model of interest.}

\item{...}{extra arguments.}

\item{pdata}{data.frame. The data used in model prediction.}
}
\value{
data.frame.
}
\description{
Persephone model prediction
}
\examples{
\dontrun{
predicted_data <- predict(object, pdata)
}
}
