% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\alias{report,ProgressModel-method}
\alias{report,ProgressModelList-method}
\title{Create reports}
\usage{
report(object, ...)

\S4method{report}{ProgressModel}(
  object,
  name = NULL,
  path = cronus::get_path_hermes(),
  test = 0.25,
  maxsam = 500,
  seed = 1
)

\S4method{report}{ProgressModelList}(object, ...)
}
\arguments{
\item{object}{an object of class \code{ProgressModel} or \code{ProgressModelList}.}

\item{...}{extra arguments.}

\item{name}{character. The html file name.}

\item{path}{character. The path to store the html report.}

\item{test}{Passed to \code{persephone::evaluate()}.}

\item{maxsam}{Passed to \code{persephone::evaluate()}.}

\item{seed}{Passed to \code{persephone::evaluate()}.}
}
\value{
nothing. The html report(s) is (are) created and stored.
}
\description{
Create html reports that present the model performance, including, fitting,
predictions, evaluation metrics and corresponding plots.
}
\examples{
\dontrun{
# Create a Region object
library(cronus)
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))

# Create a model
object1 <- new("ProgressBM",
               region = region,
               crop = "Corn",
               data = data_progress$Corn,
               formula = "CumPercentage ~ Time + agdd") # ProgressModel

# Create another model
object2 <- new("ProgressCLM",
               region = region,
               crop = "Soybeans",
               data = data_progress$Soybeans,
               formula = "Stage ~ Time + agdd + adayl") # ProgressModel

# Concatenate the models
object <- c(object1, object2) # ProgressModelList

# Fit
object <- fit(object)

# Plot
plot(object, cumulative = TRUE, seasons = 2002)

# Predict
predict(object, data_progress)

# Evaluate
object <- evaluate(object, maxsam = 100, seed = 1)
plot_metrics(object)

# Summarize
summary(object)

# Report
report(object, name = "example_report", path = getwd())
}
}
