% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmspe.R
\name{rmspe}
\alias{rmspe}
\alias{rmspe,ProgressModel-method}
\alias{rmspe,ProgressModelList-method}
\title{Root Mean Square Prediction Error}
\usage{
rmspe(object, ...)

\S4method{rmspe}{ProgressModel}(object, ...)

\S4method{rmspe}{ProgressModelList}(object, ...)
}
\arguments{
\item{object}{An object of class \code{ProgressModel} or \code{ProgressModelList}.}

\item{...}{extra arguments.}
}
\value{
A data.frame with columns Stage, Time and rmspe or an object of class
\code{ProgressModel} with its \code{metrics} slot altered.

An object identical to \code{object}, with the \code{metrics} slot(s) altered.
}
\description{
Root Mean Square Prediction Error
}
\examples{
\dontrun{
# Create a Region object
library(cronus)
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))

# Create a model
object1 <- new("ProgressBM",
               region = region,
               crop = "Corn",
               data = progress_ne$Corn,
               formula = "CumPercentage ~ Time + agdd") # ProgressModel

# Create another model
object2 <- new("ProgressCLM",
               region = region,
               crop = "Soybeans",
               data = progress_ne$Soybeans,
               formula = "Stage ~ Time + agdd + adayl") # ProgressModel

# Concatenate the models
object <- c(object1, object2) # ProgressModelList

# Fit
object <- fit(object)

# Plot
plot(object, cumulative = TRUE, seasons = 2002)

# Predict
predict(object, progress_ne)

# Evaluate
object <- crossval(object, maxsam = 100, seed = 1)
plot_metrics(object)

# Summarize
summary(object)

# Report
report(object, name = "example_report", dir = getwd())
}
}
