% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary,ProgressModelList-method}
\alias{summary,ProgressBM-method}
\alias{summary,ProgressBMM-method}
\alias{summary,ProgressCLM-method}
\title{Progress model summary}
\usage{
summary(object, ...)

\S4method{summary}{ProgressModelList}(object, ...)

\S4method{summary}{ProgressBM}(object, ...)

\S4method{summary}{ProgressBMM}(object, ...)

\S4method{summary}{ProgressCLM}(object, ...)
}
\arguments{
\item{object}{an object of class \code{ProgressModel} or \code{ProgressModelList}.}

\item{...}{extra arguments.}
}
\value{
an object of class \code{ProgressModel} or \code{ProgressModelList}.
}
\description{
Print a model summary.
}
\examples{
\dontrun{
# Create a Region object
library(cronus)
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))

# Create a model
object1 <- new("ProgressBM",
               region = region,
               crop = "Corn",
               data = progress_ne$Corn,
               formula = "CumPercentage ~ Time + agdd") # ProgressModel

# Create another model
object2 <- new("ProgressCLM",
               region = region,
               crop = "Soybeans",
               data = progress_ne$Soybeans,
               formula = "Stage ~ Time + agdd + adayl") # ProgressModel

# Concatenate the models
object <- c(object1, object2) # ProgressModelList

# Fit
object <- fit(object)

# Plot
plot(object, cumulative = TRUE, seasons = 2002)

# Predict
predict(object, progress_ne)

# Evaluate
object <- crossval(object, maxsam = 100, seed = 1)
plot_metrics(object)

# Summarize
summary(object)

# Report
report(object, name = "example_report", dir = getwd())
}
}
