% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\alias{summary,PersephoneModelList-method}
\alias{summary,PersephoneBin-method}
\alias{summary,PersephoneMixedBin-method}
\alias{summary,PersephoneCumLink-method}
\title{Persephone model summary}
\usage{
summary(object, ...)

\S4method{summary}{PersephoneModelList}(object, ...)

\S4method{summary}{PersephoneBin}(object, ...)

\S4method{summary}{PersephoneMixedBin}(object, ...)

\S4method{summary}{PersephoneCumLink}(object, ...)
}
\arguments{
\item{object}{an object of class \code{PersephoneModel} or \code{PersephoneModelList}.}

\item{...}{extra arguments.}
}
\value{
an object of class \code{PersephoneModel} or \code{PersephoneModelList}.
}
\description{
Print a model summary.
}
\examples{
\dontrun{
# Create a Region object
library(cronus)
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))

# Create a model
object1 <- new("PersephoneBin",
               region = region,
               crop = "Corn",
               data = progress_ne$Corn,
               formula = "CumPercentage ~ Time + agdd") # PersephoneModel

# Create another model
object2 <- new("PersephoneCumLink",
               region = region,
               crop = "Soybeans",
               data = progress_ne$Soybeans,
               formula = "Stage ~ Time + agdd + adayl") # PersephoneModel

# Concatenate the models
object <- c(object1, object2) # PersephoneModelList

# Fit
object <- fit(object)

# Plot
plot(object, cumulative = TRUE, seasons = 2002)

# Predict
predict(object, progress_ne)

# Evaluate
object <- crossval(object, maxsam = 100, seed = 1)
plot_metrics(object)

# Summarize
summary(object)

# Report
report(object, name = "example_report", dir = getwd())
}
}
