% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tramoseatsSingle.R
\name{tramoseatsSingle}
\alias{tramoseatsSingle}
\title{tramoseats class}
\description{
an R6 class that represents a tramoseats model
}
\section{Inherits}{
 \link{persephone}
}

\section{Contructor}{

\preformatted{
tramoseatsSingle$new(
  ts, template = c("RSAfull", "RSA0", "RSA1", "RSA2", "RSA3",
                   "RSA4", "RSA5"),
  userdefined=NULL, ...
)
}
\itemize{
\item \code{ts} denotes a time series
\item \code{template} is passed as the \code{spec} argument to \link{tramoseats_spec}
\item \code{userdefined} is passed as the \code{userdefined} argument to \link{tramoseats}
\item \code{...} passed to \link{tramoseats_spec}
}
}

\examples{
data(AirPassengers, package = "datasets")

obj <- tramoseatsSingle$new(AirPassengers, "RSA3")
obj$run()
obj$ts
}
