% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\alias{update,PersephoneModel-method}
\alias{update,PersephoneModelList-method}
\title{Update Model}
\usage{
update(object, ...)

\S4method{update}{PersephoneModel}(object, ...)

\S4method{update}{PersephoneModelList}(object, ...)
}
\arguments{
\item{object}{an object of class \code{PersephoneModel} or \code{PersephoneModelList}.
whose slots are to be updated.}

\item{...}{extra arguments.}
}
\value{
an object of class \code{PersephoneModel} or \code{PersephoneModelList}.
}
\description{
Update one or more slots of a PersephoneModel object.
}
\examples{
\dontrun{
# Create a Region object
library(cronus)
region <- Region(name = "nebraska", type = "us state",
                 div = c(country = "United States", state = "Nebraska"))

object <- new("PersephoneCumLink", region = region, crop = "Corn", model = "A")
object <- update(object, region = region, crop = "Soybeans", model = "B")
}
}
