% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\alias{update,PersephoneModel-method}
\alias{update,PersephoneModelList-method}
\title{Update Model}
\usage{
update(object, ...)

\S4method{update}{PersephoneModel}(object, ...)

\S4method{update}{PersephoneModelList}(object, ...)
}
\arguments{
\item{object}{PersephoneModel. The object whose slots are to be updated.}

\item{...}{extra arguments.}
}
\value{
PersephoneModel
}
\description{
Update one or more slots of a PersephoneModel object.
}
\examples{
\dontrun{
object <- new("PersephoneBinomial", region = "A")
object <- update(object, region = "B")
}
}
