% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x13Single.R
\name{x13Single}
\alias{x13Single}
\title{x13 class}
\description{
An R6 class that represents a x13 model.
}
\section{Inherits}{
 \link{persephone}
}

\section{Contructor}{

\preformatted{
x13Single$new(
  ts, template = c("RSA5c", "RSA0", "RSA1", "RSA2c", "RSA3", "RSA4c", "X11"), ...
)
}
\itemize{
\item \code{ts} denotes a time series
\item \code{template} is passed as the \code{spec} argument to \link{x13_spec_def}
\item \code{...} is passed to \link{x13_spec_def}
}
}

\examples{
data(AirPassengers, package = "datasets")

obj <- x13Single$new(AirPassengers, "RSA3")
obj$run()
obj$ts
}
