% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/per_x13.R
\name{x13Single}
\alias{x13Single}
\title{x13 class}
\description{
An R6 class that represents a x13 model.
}
\examples{
data(AirPassengers, package = "datasets")

obj <- per_x13(AirPassengers, "RSA3")
obj$run()
obj$ts
}
\section{Super classes}{
\code{\link[persephone:persephone]{persephone::persephone}} -> \code{\link[persephone:persephoneSingle]{persephone::persephoneSingle}} -> \code{x13Single}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-clone}{\code{x13Single$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="generate_qr_table">}\href{../../persephone/html/persephone.html#method-generate_qr_table}{\code{persephone::persephone$generate_qr_table()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="iterate">}\href{../../persephone/html/persephone.html#method-iterate}{\code{persephone::persephone$iterate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="plot">}\href{../../persephone/html/persephone.html#method-plot}{\code{persephone::persephone$plot()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="plotResiduals">}\href{../../persephone/html/persephone.html#method-plotResiduals}{\code{persephone::persephone$plotResiduals()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="plotSeasIrrCal">}\href{../../persephone/html/persephone.html#method-plotSeasIrrCal}{\code{persephone::persephone$plotSeasIrrCal()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="print">}\href{../../persephone/html/persephone.html#method-print}{\code{persephone::persephone$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="set_options">}\href{../../persephone/html/persephone.html#method-set_options}{\code{persephone::persephone$set_options()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephone" data-id="updateParams">}\href{../../persephone/html/persephone.html#method-updateParams}{\code{persephone::persephone$updateParams()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephoneSingle" data-id="initialize">}\href{../../persephone/html/persephoneSingle.html#method-initialize}{\code{persephone::persephoneSingle$initialize()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="persephone" data-topic="persephoneSingle" data-id="run">}\href{../../persephone/html/persephoneSingle.html#method-run}{\code{persephone::persephoneSingle$run()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{x13Single$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
